Ouroboros.Network.CodecCBORTerm

Documentation

data CodecCBORTerm fail a

A pure codec which encodes to / decodes from Term. This is useful if one expects a valid cbor encoding, which one might not know how to decode like in the Handshake protocol.

Constructors

CodecCBORTerm

Fields

encodeTerm ∷ a → Term

decodeTerm ∷ Term → Either fail a