Ouroboros.Network.ConnectionManager.Core

Description

The implementation of connection manager.

Documentation

data ConnectionManagerArguments handlerTrace socket peerAddr handle handleError version m

Arguments for a ConnectionManager which are independent of MuxMode.

Constructors

ConnectionManagerArguments

Fields

cmTracer ∷ Tracer m (ConnectionManagerTrace peerAddr handlerTrace)

Connection manager tracer.

cmTrTracer ∷ Tracer m (TransitionTrace peerAddr (ConnectionState peerAddr handle handleError version m))

Trace state transitions.

cmMuxTracer ∷ Tracer m (WithMuxBearer (ConnectionId peerAddr) MuxTrace)

Mux trace.

cmIPv4Address ∷ Maybe peerAddr

IPv4 address of the connection manager. If given, outbound connections to an IPv4 address will bound to it. To use bidirectional TCP connections, it must be the same as the server listening IPv4 address.

cmIPv6Address ∷ Maybe peerAddr

IPv6 address of the connection manager. If given, outbound connections to an IPv6 address will bound to it. To use bidirectional TCP connections, it must be the same as the server listening IPv6 address.

cmAddressType ∷ peerAddr → Maybe AddressType

cmSnocket ∷ Snocket m socket peerAddr

Snocket for the socket type.

cmTimeWaitTimeout ∷ DiffTime

TCP will held connections in TIME_WAIT state for up to two MSL (maximum segment time). On Linux this is set to '60' seconds on other system this might be up to four minutes.

This is configurable, so we can set different value in tests.

When this timeout expires a connection will transition from TerminatingState to TerminatedState.

cmOutboundIdleTimeout ∷ DiffTime

Inactivity timeout before the connection will be reset. It is the timeout attached to the OutboundIdleState.

connectionDataFlow ∷ version → DataFlow

version represents the tuple of versionNumber and agreedOptions.

cmPrunePolicy ∷ PrunePolicy peerAddr (STM m)

Prune policy

cmConnectionsLimits ∷ AcceptedConnectionsLimit

withConnectionManager

Arguments

∷ ∀ (muxMode ∷ MuxMode) peerAddr socket handlerTrace handle handleError version m a. (Monad m, MonadLabelledSTM m, MonadFork m, MonadAsync m, MonadEvaluate m, MonadMask m, MonadMonotonicTime m, MonadThrow (STM m), MonadTimer m, Ord peerAddr, Show peerAddr, Typeable peerAddr)

⇒ ConnectionManagerArguments handlerTrace socket peerAddr handle handleError version m

→ ConnectionHandler muxMode handlerTrace socket peerAddr handle handleError version m
Callback which runs in a thread dedicated for a given connection.

→ (handleError → HandleErrorType)
classify handleErrors

→ InResponderMode muxMode (ControlChannel m (NewConnection peerAddr handle))
On outbound duplex connections we need to notify the server about a new connection.

→ (ConnectionManager muxMode socket peerAddr handle handleError m → m a)
Continuation which receives the ConnectionManager. It must not leak outside of scope of this callback. Once it returns all resources will be closed.

→ m a

Entry point for using the connection manager. This is a classic with style combinator, which cleans resources on exit of the callback (whether cleanly or through an exception).

Including a connection (either inbound or outbound) is an idempotent operation on connection manager state. The connection manager will always return the handle that was first to be included in its state.

Once an inbound connection is passed to the ConnectionManager, the manager is responsible for the resource.

defaultTimeWaitTimeout ∷ DiffTime

The default value for cmTimeWaitTimeout.

defaultProtocolIdleTimeout ∷ DiffTime

Inactivity timeout. It configures how long to wait since the local side demoted remote peer to cold, before closing the connection.

defaultResetTimeout ∷ DiffTime

data ConnectionState peerAddr handle handleError version m

State of a connection.

Constructors

ReservedOutboundState
Each outbound connections starts in this state.

UnnegotiatedState !Provenance !(ConnectionId peerAddr) !(Async m ())
Each inbound connection starts in this state, outbound connection reach this state once connect call returns.

OutboundUniState !(ConnectionId peerAddr) !(Async m ()) !handle
OutboundState Unidirectional state.

OutboundDupState !(ConnectionId peerAddr) !(Async m ()) !handle !TimeoutExpired
Either OutboundState Duplex or OutobundState^tau Duplex.

OutboundIdleState !(ConnectionId peerAddr) !(Async m ()) !handle !DataFlow
Before connection is reset it is put in OutboundIdleState for the duration of cmOutboundIdleTimeout.

InboundIdleState !(ConnectionId peerAddr) !(Async m ()) !handle !DataFlow

InboundState !(ConnectionId peerAddr) !(Async m ()) !handle !DataFlow

DuplexState !(ConnectionId peerAddr) !(Async m ()) !handle

TerminatingState !(ConnectionId peerAddr) !(Async m ()) !(Maybe handleError)

TerminatedState !(Maybe handleError)

abstractState ∷ MaybeUnknown (ConnectionState muxMode peerAddr m a b) → AbstractState