Ouroboros.Network.ConnectionManager.Types

Description

Connection manager core types.

Connection manager is responsible for managing uni- and bi-directional connections and threads which are running network applications using 'network-mux'. In particular it is responsible for:

opening new connection / reusing connections (for bidirectional connections) and exposes a method to register inbound connections;

run connection handler, i.e. ConnectionHandler, which runs handshake negotiation, notifies connection manager on the results and starts the multiplexer;

error handling for connection threads;

keeping track of handshake negotiation: whether a unidirectional or duplex connection was negotiated;

tracking state of each connection;

keep inbound connections under limits.

Connection manager is designed to work for any MuxMode, though the most useful ones are ResponderMode and InitiatorResponderMode:

InitiatorResponderMode - useful for node-to-node applications, which needs to create outbound connections as well as accept inbound ones;

ResponderMode - useful for server side of node-to-client; it allows us to share the same server between node-to-client and node-to-node;

InitiatorMode - could be used on client side of node-to-client applications.

The calls requestOutboundConnection and includeInboundConnection return once a connection has been negotiated. The returned handle contains all the information that is need to start and monitor mini-protocols through the mux interface.

For inbound connections, the connection manager will pass handle (also after negotiation).

┌────────────────────────┐
│                        │        ┏━━━━━━━━━━━━━━━━┓     ┌──────────────────┐
│   ConnectionHandler    │        ┃                ┃     │                  │
│                        ┝━━━━━━━▶┃     handle     ┣━━━━▶│ PeerStateActions ├───▶ P2P Governor
│  inbound / outbound    │        ┃                ┃     │                  │
│         ┃              │        ┗━━┳━━━━━━━━━━━━━┛     └──────────────────┘
└─────────╂──────────────┘           ┃
          ┃                          ┃
          ▼                          ┃
   ┏━━━━━━━━━━━━━━━━━┓               ┃
   ┃ Control Channel ┃               ┃
   ┗━━━━━━┳━━━━━━━━━━┛               ┃
          ┃                          ┃
          ┃                          ┃
          ▼                          ┃
┌────────────────────────┐           ┃
│                        │           ┃
│         Server         │◀━━━━━━━━━━┛
│                        │
└────────────────────────┘

Termination procedure as well as connection state machine is not described in this haddock, see associated specification.

Connection manager core types

Connection Types

data AddressType

Connection manager supports IPv4 and IPv6 addresses.

Constructors

IPv4Address

IPv6Address

data Provenance

Each connection is is either initiated locally (outbound) or by a remote peer (inbound).

Constructors

Inbound
An inbound connection: one that was initiated by a remote peer.

Outbound
An outbound connection: one that was initiated by us.

data DataFlow

Each connection negotiates if it is uni- or bi-directional. DataFlow is a life time property of a connection, once negotiated it never changes.

Constructors

Unidirectional

Duplex

data TimeoutExpired

Boolean like type which indicates if the timeout on OutboundStateDuplex has expired.

Constructors

Expired

Ticking

data ConnectionType

Either unnegotiated or negotiated unidirectional or duplex connections. This is not a static property of a connection. It is used by PrunePolicy.

Note: the order matters, it can be used by a PickPolicy, e.g. simplePickPolicy.

Constructors

UnnegotiatedConn !Provenance
An unnegotiated connection.

InboundIdleConn !DataFlow
An inbound idle connection.

OutboundIdleConn !DataFlow
An outbound idle connection.

NegotiatedConn !Provenance !DataFlow
A negotiated connection, which is used in only one direction indicated by Provenance. The connection could itself negotiated either Duplex or Unidirectional data flow.

DuplexConn
A connection which is running in full duplex mode.

Connection Handler

ConnectionHandler provides monadic action which runs handshake negotiation and starts the multiplexer. It's the component which has access to underlying socket. There's one-to-one correspondence between sockets and threads that run the handler.

ConnectionHandlerFn

is the type of callback executed for each connection. All arguments are provided by the connection manager.

ConnectionHandler

is a newtype wrapper which provides inbound / outbound handlers depending on MuxMode.

newtype MaskedAction m a

Handler action is started with asynchronous exceptions masked; this allows to install exception handlers in an async-safe way.

Constructors

MaskedAction

Fields

runWithUnmask ∷ (∀ x. m x → m x) → m a

type ConnectionHandlerFn handlerTrace socket peerAddr handle handleError version m = socket → PromiseWriter m (Either handleError (handle, version)) → Tracer m handlerTrace → ConnectionId peerAddr → (DiffTime → socket → m (MuxBearer m)) → MaskedAction m ()

MaskedAction which is executed by thread designated for a given connection.

PromiseWriter allows to notify the ConnectionManager about the result of handshake negotiation.

Note: PromiseWriter could be replaced with an stm action which is accessing the TVar which holds state of the connection.

data ConnectionHandler muxMode handlerTrace socket peerAddr handle handleError version m

Connection handler action. It is index by muxMode :: MuxMode. There's one ConnectionHandlerFn per provenance, possibly limited by muxMode.

Constructors

ConnectionHandler

Fields

connectionHandler ∷ WithMuxTuple muxMode (ConnectionHandlerFn handlerTrace socket peerAddr handle handleError version m)

Connection handler.

data Inactive

Boolean like type

Constructors

Active MiniProtocolDir

Inactive

data ExceptionInHandler peerAddr where

Exception which where caught in the connection thread and were re-thrown in the main thread by the rethrowPolicy.

Constructors

ExceptionInHandler ∷ !peerAddr → !SomeException → ExceptionInHandler peerAddr

data HandleErrorType

Data type used to classify handleErrors.

Constructors

HandshakeFailure
Handshake negotiation failed. This is not a protocol error.

HandshakeProtocolViolation
Handshake protocol error. This should include timeout errors or any IO errors.

Prune Policy

type PrunePolicy peerAddr stm = Map peerAddr ConnectionType → Int → stm (Set peerAddr)

PrunePolicy allows to pick a select peers from which we will disconnect (we use TCP reset). The chosen connections will be terminated by the connection manger once it detects that there are too many inbound connections.

simplePrunePolicy ∷ (Applicative stm, Ord peerAddr) ⇒ PrunePolicy peerAddr stm

The simplest PrunePolicy, it should only be used for tests.

Connection Manager

Connection Manager Arguments

data ConnectionManager (muxMode ∷ MuxMode) socket peerAddr handle handleError m

ConnectionManager.

We identify resources (e.g. Socket or HANDLE) by their address. It is enough for us to use just the remote address rather than connection identifier, since we need one connection towards that peer, even if we are connected through multiple local addresses. It is safe to share a connection manager with multiple listening sockets.

Constructors

ConnectionManager

Fields

getConnectionManager ∷ WithMuxMode muxMode (OutboundConnectionManager muxMode socket peerAddr handle handleError m) (InboundConnectionManager muxMode socket peerAddr handle handleError m)

readState ∷ m (Map peerAddr AbstractState)

API

data Connected peerAddr handle handleError

Result of requestOutboundConnection or includeInboundConnection.

Constructors

Connected !(ConnectionId peerAddr) !DataFlow !handle
We are connected and mux is running.

Disconnected !(ConnectionId peerAddr) !(Maybe handleError)
There was an error during handshake negotiation.

Implementation detail: we return Maybe handleError, rather than handleError. In case of an existing inbound connection, the implementation of requestOutboundConnection is awaiting on handshake through the connection state. The TerminatingState or TerminatedState are not only used for handshake errors, but also for normal termination, hence the Maybe. We could await on update from the handshake instead, but this would introduce a race between inbound / outbound threads.

data OperationResult a

Custom either type for result of various methods.

Constructors

UnsupportedState !AbstractState

OperationSuccess !a

TerminatedConnection !AbstractState

resultInState ∷ OperationResult AbstractState → AbstractState

data DemotedToColdRemoteTr

Return value of unregisterInboundConnection to inform the caller about the transition.

Constructors

CommitTr
Commit^{dataFlow} transition from InboundIdleState dataFlow.

KeepTr
Either DemotedToCold^{Remote} transition from DuplexState, or a level triggered Awake^{Duplex}_{Local} transition. In both cases the server must keep the responder side of all protocols ready.

type RequestOutboundConnection peerAddr handle handleError m = peerAddr → m (Connected peerAddr handle handleError)

type IncludeInboundConnection socket peerAddr handle handleError m

Arguments

= Word32
inbound connections hard limit. NOTE: Check TODO over at includeInboundConnectionImpl definition.

→ socket

→ peerAddr

→ m (Connected peerAddr handle handleError)

Outbound side

requestOutboundConnection ∷ HasInitiator muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → RequestOutboundConnection peerAddr handle handleError m

Include outbound connection into ConnectionManager.

This executes:

\(Reserve\) to \(Negotiated^{*}_{Outbound}\) transitions

\(PromotedToWarm^{Duplex}_{Local}\) transition

\(Awake^{Duplex}_{Local}\) transition

promotedToWarmRemote ∷ HasResponder muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → peerAddr → m (OperationResult AbstractState)

Notify the ConnectionManager that a remote end promoted us to a warm peer.

This executes either:

\(PromotedToWarm^{Duplex}_{Remote}\) transition,

\(Awake^{*}_{Remote}\) transition

from the specification.

demotedToColdRemote ∷ HasResponder muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → peerAddr → m (OperationResult AbstractState)

Notify the ConnectionManager that a remote end demoted us to a /cold peer/.

This executes:

\(DemotedToCold^{*}_{Remote}\) transition.

This method is idempotent.

unregisterOutboundConnection

Arguments

∷ HasInitiator muxMode ~ True

⇒ ConnectionManager muxMode socket peerAddr handle handleError m

→ peerAddr

→ m (OperationResult AbstractState)
reports the from-state.

Unregister outbound connection.

This executes:

\(DemotedToCold^{*}_{Local}\) transitions

Inbound side

includeInboundConnection ∷ HasResponder muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → IncludeInboundConnection socket peerAddr handle handleError m

Include an inbound connection into ConnectionManager. This executes:

\(Reserve\) to \(Negotiated^{*}_{Outbound}\) transitions

\(PromotedToWarm^{Duplex}_{Local}\) transition

\(Awake^{Duplex}_{Local}\) transition

unregisterInboundConnection ∷ HasResponder muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → peerAddr → m (OperationResult DemotedToColdRemoteTr)

Unregister outbound connection. Returns if the operation was successul.

This executes:

\(Commit*{*}\) transition

\(TimeoutExpired\) transition

numberOfConnections ∷ HasResponder muxMode ~ True ⇒ ConnectionManager muxMode socket peerAddr handle handleError m → STM m Int

Number of connections tracked by the server.

Private API

data OutboundConnectionManager (muxMode ∷ MuxMode) socket peerAddr handle handleError m where

Outbound connection manager API.

Constructors

OutboundConnectionManager

Fields

∷ HasInitiator muxMode ~ True

⇒ { ocmRequestConnection ∷ RequestOutboundConnection peerAddr handle handleError m

, ocmUnregisterConnection ∷ peerAddr → m (OperationResult AbstractState)

} → OutboundConnectionManager muxMode socket peerAddr handle handleError m

data InboundConnectionManager (muxMode ∷ MuxMode) socket peerAddr handle handleError m where

Inbound connection manager API. For a server implementation we also need to know how many connections are now managed by the connection manager.

This type is an internal detail of ConnectionManager

Constructors

InboundConnectionManager

Fields

∷ HasResponder muxMode ~ True

⇒ { icmIncludeConnection ∷ IncludeInboundConnection socket peerAddr handle handleError m

, icmUnregisterConnection ∷ peerAddr → m (OperationResult DemotedToColdRemoteTr)

, icmPromotedToWarmRemote ∷ peerAddr → m (OperationResult AbstractState)

, icmDemotedToColdRemote ∷ peerAddr → m (OperationResult AbstractState)

, icmNumberOfConnections ∷ STM m Int

} → InboundConnectionManager muxMode socket peerAddr handle handleError m

Exceptions

data ConnectionManagerError peerAddr

Exceptions used by ConnectionManager.

Constructors

ConnectionExists !Provenance !peerAddr !CallStack
A connection manager was asked for an outbound connection and there either exists a connection used in outbound direction or a reservation for an outbound connection.

ForbiddenConnection !(ConnectionId peerAddr) !CallStack
Connection manager was asked for an outbound connection which turned out to be unidirectional inbound, and thus it cannot be re-used..

ImpossibleConnection !(ConnectionId peerAddr) !CallStack
Connections that would be forbidden by the kernel (TCP semantics).

ConnectionTerminating !(ConnectionId peerAddr) !CallStack
Connection is now terminating.

ConnectionTerminated !peerAddr !CallStack
Connection has terminated.

ImpossibleState !peerAddr !CallStack
Connection manager in impossible state.

ForbiddenOperation !peerAddr !AbstractState !CallStack
A forbidden operation in the given connection state.

UnknownPeer !peerAddr !CallStack
A connection does not exists. Only thrown when an existing connection was expected.

data SomeConnectionManagerError

Existential wrapper for ConnectionManagerError peerAddr. It allows to use fromException, without being bothered about the address type.

Constructors

∀ addr.(Typeable addr, Show addr) ⇒ SomeConnectionManagerError !(ConnectionManagerError addr)

data AbstractState

Useful for tracing and error messages.

Constructors

UnknownConnectionSt

ReservedOutboundSt

UnnegotiatedSt !Provenance

InboundIdleSt !DataFlow

InboundSt !DataFlow

OutboundUniSt

OutboundDupSt !TimeoutExpired

OutboundIdleSt !DataFlow

DuplexSt

WaitRemoteIdleSt

TerminatingSt

TerminatedSt

Counters

data ConnectionManagerCounters

Counters for tracing and analysis purposes

Constructors

ConnectionManagerCounters

Fields

fullDuplexConns ∷ !Int

number of full duplex connections

duplexConns ∷ !Int

number of negotiated duplex connections (including DuplexState connections)

unidirectionalConns ∷ !Int

number of negotiated unidirectional connections

inboundConns ∷ !Int

number of inbound connections

outboundConns ∷ !Int

number of outbound connections

Mux types

data WithMuxMode (mode ∷ MuxMode) a b where

Constructors

WithInitiatorMode ∷ a → WithMuxMode InitiatorMode a b

WithResponderMode ∷ b → WithMuxMode ResponderMode a b

WithInitiatorResponderMode ∷ a → b → WithMuxMode InitiatorResponderMode a b

withInitiatorMode ∷ HasInitiator mode ~ True ⇒ WithMuxMode mode a b → a

withResponderMode ∷ HasResponder mode ~ True ⇒ WithMuxMode mode a b → b

Promise

Promise interface, backed by a StrictTMVar.

Making two separate interfaces: PromiseWriter and PromiseReader allows us to make a clear distinction between consumer and producers threads.

newEmptyPromiseIO ∷ (MonadSTM m, MonadThrow (STM m)) ⇒ m (PromiseReader m a, PromiseWriter m a)

newtype PromiseReader m a

Constructors

PromiseReader

Fields

readPromise ∷ STM m a

A blocking read operation.

readPromiseIO ∷ MonadSTM m ⇒ PromiseReader m a → m a

data PromiseWriter m a

Constructors

PromiseWriter

Fields

writePromise ∷ a → STM m ()

putPromise, is a non-blocking operation, it throws PromiseWriterException if it would block.

forcePromise ∷ a → STM m ()

If the promise is empty it fills it, if it is non-empty it replaces the current value.

data PromiseWriterException

Constructors

PromiseWriterBlocked

Tracing

data AssertionLocation peerAddr

AssertionLocation contains constructors that situate the location of the tracing so one can be sure where the assertion came from as well as the all relevant information.

Constructors

ForkConnectionHandlerCleanup !(Maybe (ConnectionId peerAddr)) !AbstractState

UnregisterInboundConnection !(Maybe (ConnectionId peerAddr)) !AbstractState

RequestOutboundConnection !(Maybe (ConnectionId peerAddr)) !AbstractState

UnregisterOutboundConnection !(Maybe (ConnectionId peerAddr)) !AbstractState

PromotedToWarmRemote !(Maybe (ConnectionId peerAddr)) !AbstractState

DemotedToColdRemote !(Maybe (ConnectionId peerAddr)) !AbstractState

data ConnectionManagerTrace peerAddr handlerTrace

ConnectionManagerTrace contains a hole for a trace of single connection which is filled with ConnectionHandlerTrace.

Constructors

TrIncludeConnection Provenance peerAddr

TrUnregisterConnection Provenance peerAddr

TrConnect

Fields

(Maybe peerAddr)

local address

peerAddr

remote address

TrConnectError

Fields

(Maybe peerAddr)

local address

peerAddr

remote address

SomeException

TrTerminatingConnection Provenance (ConnectionId peerAddr)

TrTerminatedConnection Provenance peerAddr

TrConnectionHandler (ConnectionId peerAddr) handlerTrace

TrShutdown

TrConnectionExists Provenance peerAddr AbstractState

TrForbiddenConnection (ConnectionId peerAddr)

TrImpossibleConnection (ConnectionId peerAddr)

TrConnectionFailure (ConnectionId peerAddr)

TrConnectionNotFound Provenance peerAddr

TrForbiddenOperation peerAddr AbstractState

TrPruneConnections

Fields

(Set peerAddr)

prunning set

Int

number connections that must be prunned

(Set peerAddr)

choice set

TrConnectionCleanup (ConnectionId peerAddr)

TrConnectionTimeWait (ConnectionId peerAddr)

TrConnectionTimeWaitDone (ConnectionId peerAddr)

TrConnectionManagerCounters ConnectionManagerCounters

TrState (Map peerAddr AbstractState)
traced on SIGUSR1 signal, installed in runDataDiffusion

TrUnexpectedlyFalseAssertion (AssertionLocation peerAddr)
This case is unexpected at call site.

data MaybeUnknown state

A custom version of Maybe type, which allows to explicitly represent connections which are not registered by the connection manager.

Constructors

Known !state
Known connection in state

Race !state
There is a possible race condition between connection finalizer and either inbound or outbound connection registration. If that happens we use Race constructor.

Unknown
Connection is is not known to the connection manager.

data Transition' state

Constructors

Transition

Fields

fromState ∷ !state

toState ∷ !state

type Transition state = Transition' (MaybeUnknown state)

type AbstractTransition = Transition' AbstractState

mkTransition ∷ state → state → Transition state

type TransitionTrace peerAddr state = TransitionTrace' peerAddr (MaybeUnknown state)

data TransitionTrace' peerAddr state

Constructors

TransitionTrace

Fields

ttPeerAddr ∷ peerAddr

ttTransition ∷ Transition' state

type AbstractTransitionTrace peerAddr = TransitionTrace' peerAddr AbstractState