Ouroboros.Network.Driver.Limits

Description

Drivers for running Peers.

Limits

data ProtocolSizeLimits ps bytes

Constructors

ProtocolSizeLimits

Fields

sizeLimitForState ∷ ∀ (pr ∷ PeerRole) (st ∷ ps). PeerHasAgency pr st → Word

dataSize ∷ bytes → Word

data ProtocolTimeLimits ps

Constructors

ProtocolTimeLimits

Fields

timeLimitForState ∷ ∀ (pr ∷ PeerRole) (st ∷ ps). PeerHasAgency pr st → Maybe DiffTime

data ProtocolLimitFailure

Constructors

ExceededSizeLimit

ExceededTimeLimit

Normal peers

runPeerWithLimits ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadTimer m, MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → ProtocolSizeLimits ps bytes → ProtocolTimeLimits ps → Channel m bytes → Peer ps pr st m a → m a

data TraceSendRecv ps where

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg ∷ AnyMessage ps → TraceSendRecv ps

TraceRecvMsg ∷ AnyMessage ps → TraceSendRecv ps

Pipelined peers

runPipelinedPeerWithLimits ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadAsync m, MonadTimer m, MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → ProtocolSizeLimits ps bytes → ProtocolTimeLimits ps → Channel m bytes → PeerPipelined ps pr st m a → m a

Run a pipelined peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

Unlike normal peers, running pipelined peers rely on concurrency, hence the MonadSTM constraint.

Driver utilities

driverWithLimits ∷ ∀ ps failure bytes m. (MonadThrow m, MonadTimer m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → ProtocolSizeLimits ps bytes → ProtocolTimeLimits ps → Channel m bytes → Driver ps (Maybe bytes) m