Ouroboros.Network.Driver.Simple

Description

Drivers for running Peers with a Codec and a Channel.

Introduction

A Peer is a particular implementation of an agent that engages in a typed protocol. To actualy run one we need a source and sink for the typed protocol messages. These are provided by a Channel and a Codec. The Channel represents one end of an untyped duplex message transport, and the Codec handles conversion between the typed protocol messages and the untyped channel.

So given the Peer and a compatible Codec and Channel we can run the peer in some appropriate monad. The peer and codec have to agree on the same protocol and role in that protocol. The codec and channel have to agree on the same untyped medium, e.g. text or bytes. All three have to agree on the same monad in which they will run.

This module provides drivers for normal and pipelined peers. There is very little policy involved here so typically it should be possible to use these drivers, and customise things by adjusting the peer, or codec or channel.

It is of course possible to write custom drivers and the code for these ones may provide a useful starting point. The runDecoder function may be a helpful utility for use in custom drives.

Normal peers

runPeer ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Channel m bytes → Peer ps pr st m a → m a

Run a peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

data TraceSendRecv ps where

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg ∷ AnyMessage ps → TraceSendRecv ps

TraceRecvMsg ∷ AnyMessage ps → TraceSendRecv ps

Pipelined peers

runPipelinedPeer ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadSTM m, MonadAsync m, MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Channel m bytes → PeerPipelined ps pr st m a → m a

Run a pipelined peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

Unlike normal peers, running pipelined peers rely on concurrency, hence the MonadSTM constraint.

Connected peers

runConnectedPeers ∷ (MonadSTM m, MonadAsync m, MonadCatch m, Exception failure) ⇒ m (Channel m bytes, Channel m bytes) → Tracer m (PeerRole, TraceSendRecv ps) → Codec ps failure m bytes → Peer ps pr st m a → Peer ps (FlipAgency pr) st m b → m (a, b)

Run two Peers via a pair of connected Channels and a common Codec.

This is useful for tests and quick experiments.

The first argument is expected to create two channels that are connected, for example createConnectedChannels.

runConnectedPeersPipelined ∷ (MonadSTM m, MonadAsync m, MonadCatch m, Exception failure) ⇒ m (Channel m bytes, Channel m bytes) → Tracer m (PeerRole, TraceSendRecv ps) → Codec ps failure m bytes → PeerPipelined ps pr st m a → Peer ps (FlipAgency pr) st m b → m (a, b)