Ouroboros.Network.ErrorPolicy

Description

Error policies, and integration with SuspendDecision-semigroup action on PeerState.

Documentation

data ErrorPolicies

List of error policies for exception handling and a policy for handing application return values.

Constructors

ErrorPolicies

Fields

epAppErrorPolicies ∷ [ErrorPolicy]

Application Error Policies

epConErrorPolicies ∷ [ErrorPolicy]

connect Error Policies

nullErrorPolicies ∷ ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy

Fields

∷ Exception e

⇒ (e → Maybe (SuspendDecision DiffTime))

Nothing means no decision. It is equivalent to not having the policy at all. In evalErrorPolicies this will fall-through and match against the remaining policies.

→ ErrorPolicy

evalErrorPolicy ∷ ∀ e. Exception e ⇒ e → ErrorPolicy → Maybe (SuspendDecision DiffTime)

evalErrorPolicies ∷ ∀ e. Exception e ⇒ e → [ErrorPolicy] → Maybe (SuspendDecision DiffTime)

Evaluate a list of ErrorPolicys; If none of them applies this function returns Nothing, in this case the exception will be traced and not thrown.

type CompleteApplication m s addr r = Result addr r → s → STM m (CompleteApplicationResult m addr s)

Complete a connection, which receive application result (or exception).

data CompleteApplicationResult m addr s

Constructors

CompleteApplicationResult

Fields

carState ∷ !s

new state

carThreads ∷ Set (Async m ())

threads to kill

carTrace ∷ Maybe (WithAddr addr ErrorPolicyTrace)

trace points

data Result addr r where

Result of the connection thread. It's either result of an application, or an exception thrown by it.

Constructors

ApplicationResult ∷ !Time → !addr → !r → Result addr r

Connected ∷ !Time → !addr → Result addr r

ConnectionError ∷ Exception e ⇒ !Time → !addr → !e → Result addr r

ApplicationError ∷ Exception e ⇒ !Time → !addr → !e → Result addr r

completeApplicationTx ∷ ∀ m addr a. (MonadAsync m, Ord addr, Ord (Async m ())) ⇒ ErrorPolicies → CompleteApplication m (PeerStates m addr) addr a

CompleteApplication callback

Traces

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime !DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

traceErrorPolicy ∷ Either (ConnectionOrApplicationExceptionTrace SomeException) r → SuspendDecision DiffTime → Maybe ErrorPolicyTrace

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr ∷ !addr

wiaEvent ∷ !a

Re-exports of PeerState

data PeerStates m addr

Map from addresses to PeerStates; it will be be shared in a StrictTVar.

Abstracting t is useful for tests, the IO version will use Time IO.

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.