Ouroboros.Network.InboundGovernor

Description

Server implementation based on ConnectionManager

Documentation

newtype InboundGovernorObservableState

Currently only StdGen, but in the future this will be extended to a record which contains some useful statistics about peers to support more advances prune strategies (see. PruneStrategy).

Constructors

InboundGovernorObservableState

Fields

igosPrng ∷ StdGen

newObservableStateVar ∷ MonadLabelledSTM m ⇒ StdGen → m (StrictTVar m InboundGovernorObservableState)

Create new observable state StrictTVar.

newObservableStateVarIO ∷ IO (StrictTVar IO InboundGovernorObservableState)

Using the global StdGen.

newObservableStateVarFromSeed ∷ MonadLabelledSTM m ⇒ Int → m (StrictTVar m InboundGovernorObservableState)

Useful for testing, it is using mkStdGen.

Run Inbound Protocol Governor

inboundGovernor ∷ ∀ (muxMode ∷ MuxMode) socket peerAddr versionNumber m a b. (MonadAsync m, MonadCatch m, MonadEvaluate m, MonadThrow m, MonadThrow (STM m), MonadTime m, MonadTimer m, MonadMask m, Ord peerAddr, HasResponder muxMode ~ True) ⇒ Tracer m (RemoteTransitionTrace peerAddr) → Tracer m (InboundGovernorTrace peerAddr) → ServerControlChannel muxMode peerAddr ByteString m a b → DiffTime → MuxConnectionManager muxMode socket peerAddr versionNumber ByteString m a b → StrictTVar m InboundGovernorObservableState → m Void

Run the server, which consists of the following components:

inbound governor, it corresponds to p2p-governor on outbound side

accept loop(s), one per given ip address. We support up to one ipv4 address and up to one ipv6 address, i.e. an ipv6 enabled node will run two accept loops on listening on different addresses with shared inbound governor.

The server can be run in either of two MuxMode-es:

InitiatorResponderMode

ResponderMode

The first one is used in data diffusion for Node-To-Node protocol, while the other is useful for running a server for the Node-To-Client protocol.

Trace

data InboundGovernorTrace peerAddr

Constructors

TrNewConnection !Provenance !(ConnectionId peerAddr)

TrResponderRestarted !(ConnectionId peerAddr) !MiniProtocolNum

TrResponderStartFailure !(ConnectionId peerAddr) !MiniProtocolNum !SomeException

TrResponderErrored !(ConnectionId peerAddr) !MiniProtocolNum !SomeException

TrResponderStarted !(ConnectionId peerAddr) !MiniProtocolNum

TrResponderTerminated !(ConnectionId peerAddr) !MiniProtocolNum

TrPromotedToWarmRemote !(ConnectionId peerAddr) !(OperationResult AbstractState)

TrPromotedToHotRemote !(ConnectionId peerAddr)

TrDemotedToWarmRemote !(ConnectionId peerAddr)

TrDemotedToColdRemote !(ConnectionId peerAddr) !(OperationResult DemotedToColdRemoteTr)
All mini-protocols terminated. The boolean is true if this connection was not used by p2p-governor, and thus the connection will be terminated.

TrWaitIdleRemote !(ConnectionId peerAddr) !(OperationResult AbstractState)

TrMuxCleanExit !(ConnectionId peerAddr)

TrMuxErrored !(ConnectionId peerAddr) SomeException

TrInboundGovernorCounters !InboundGovernorCounters

TrRemoteState !(Map (ConnectionId peerAddr) RemoteSt)

TrUnexpectedlyFalseAssertion !(IGAssertionLocation peerAddr)
This case is unexpected at call site.

TrInboundGovernorError !SomeException

data RemoteSt

Remote connection state tracked by inbound protocol governor.

Constructors

RemoteWarmSt

RemoteHotSt

RemoteIdleSt

RemoteColdSt

type RemoteTransition = Transition' (Maybe RemoteSt)

Nothing represents unitialised state.

type RemoteTransitionTrace peerAddr = TransitionTrace' peerAddr (Maybe RemoteSt)

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

ServerTraceAcceptConnectionResume Int

Re-exports

data Transition' state

Constructors

Transition

Fields

fromState ∷ !state

toState ∷ !state

data TransitionTrace' peerAddr state

Constructors

TransitionTrace

Fields

ttPeerAddr ∷ peerAddr

ttTransition ∷ Transition' state