Ouroboros.Network.InboundGovernor.ControlChannel

Description

Intended to be imported qualified.

Documentation

data NewConnection peerAddr handle

Announcement message for a new connection.

Constructors

NewConnection !Provenance !(ConnectionId peerAddr) !DataFlow !handle
Announce a new connection. Inbound protocol governor will start responder protocols using StartOnDemand strategy and monitor remote transitions: PromotedToWarm^{Duplex}_{Remote} and DemotedToCold^{dataFlow}_{Remote}.

data ControlChannel peerAddr handle m

Control channel. It allows to pass STM transactions which will resolve to NewConnection. Server's monitoring thread is the consumer of these messages; there are two producers: accept loop and connection handler for outbound connections.

Constructors

ControlChannel

Fields

readMessage ∷ STM m (NewConnection peerAddr handle)

Read a single NewConnection instruction from the channel.

writeMessage ∷ NewConnection peerAddr handle → STM m ()

Write a NewConnection to the channel.

type ServerControlChannel (muxMode ∷ MuxMode) peerAddr bytes m a b = ControlChannel peerAddr (Handle muxMode peerAddr bytes m a b) m

A Server control channel which instantiates handle.

newControlChannel ∷ ∀ peerAddr handle m. MonadLabelledSTM m ⇒ m (ControlChannel peerAddr handle m)