Ouroboros.Network.InboundGovernor.State

Documentation

newtype InboundGovernorObservableState

Currently only StdGen, but in the future this will be extended to a record which contains some useful statistics about peers to support more advances prune strategies (see. PruneStrategy).

Constructors

InboundGovernorObservableState

Fields

igosPrng ∷ StdGen

newObservableStateVar ∷ MonadLabelledSTM m ⇒ StdGen → m (StrictTVar m InboundGovernorObservableState)

Create new observable state StrictTVar.

newObservableStateVarIO ∷ IO (StrictTVar IO InboundGovernorObservableState)

Using the global StdGen.

newObservableStateVarFromSeed ∷ MonadLabelledSTM m ⇒ Int → m (StrictTVar m InboundGovernorObservableState)

Useful for testing, it is using mkStdGen.

Internals

data InboundGovernorState muxMode peerAddr m a b

InboundGovernorState, which consist of pure part, and a mutable part. The mutable part can be observable from outside. Future version could contain additional statistics on the peers.

Constructors

InboundGovernorState

Fields

igsConnections ∷ !(Map (ConnectionId peerAddr) (ConnectionState muxMode peerAddr m a b))

Map of connections state. Modifying igsConnections outside of inboundGovernorLoop is not safe.

igsObservableVar ∷ !(StrictTVar m InboundGovernorObservableState)

PRNG available to PrunePolicy.

igsCountersCache ∷ !(Cache InboundGovernorCounters)

InboundGovernorCounters counters cache. Allows to only trace values when necessary.

data ConnectionState muxMode peerAddr m a b

Per connection state tracked by inbound protocol governor.

Constructors

ConnectionState

Fields

csMux ∷ !(Mux muxMode m)

Mux interface.

csDataFlow ∷ !DataFlow

Connection data flow.

csMiniProtocolMap ∷ !(Map MiniProtocolNum (MiniProtocolData muxMode m a b))

All supported mini-protocols and respective ProtocolTemperature

csCompletionMap ∷ !(Map MiniProtocolNum (STM m (Either SomeException b)))

Map of all running mini-protocol completion STM actions.

csRemoteState ∷ !(RemoteState m)

State of the connection.

data InboundGovernorCounters

Counters for tracing and analysis purposes

Constructors

InboundGovernorCounters

Fields

coldPeersRemote ∷ !Int

the number of remote peers which are in RemoteCold state

idlePeersRemote ∷ !Int

the number of remote peers which are in RemoteIdle state

warmPeersRemote ∷ !Int

the number of remote peers which are in RemoteWarm state (a close approximation of peers that have the node as a warm peer)

hotPeersRemote ∷ !Int

the number of remote peers which are in RemoteHot state (a close approximation of peers that have the node as a hot peer)

inboundGovernorCounters ∷ InboundGovernorState muxMode peerAddr m a b → InboundGovernorCounters

unregisterConnection ∷ Ord peerAddr ⇒ ConnectionId peerAddr → InboundGovernorState muxMode peerAddr m a b → InboundGovernorState muxMode peerAddr m a b

Remove connection from InboundGovernorState.

updateMiniProtocol ∷ Ord peerAddr ⇒ ConnectionId peerAddr → MiniProtocolNum → STM m (Either SomeException b) → InboundGovernorState muxMode peerAddr m a b → InboundGovernorState muxMode peerAddr m a b

Update a mini-protocol in ConnectionState. Once a mini-protocol was restarted we put the new completion action into csCompletionMap.

data RemoteState m

Each inbound connection is either in RemoteIdle, RemoteCold or RemoteEstablished state. We only need to support PromotedToWarm^{Duplex}_{Remote}, DemotedToCold^{Duplex}_{Remote} and DemotedToCold^{Unidirectional}_{Remote} transitions.

Constructors

RemoteWarm
After PromotedToWarm^{dataFlow}_{Remote} a connection is in RemoteWarm state.

RemoteHot
In this state all established and hot mini-protocols are running and none of the warm mini-protocols is running.

RemoteIdle !(STM m ())
After DemotedToCold^{dataFlow}_{Remote} is detected. This state corresponds to InboundIdleState. In this state we are checking if the responder protocols are idle during protocol idle timeout (represented by an STM action)

RemoteIdle is the initial state of an accepted a connection.

RemoteCold
The RemoteCold state for Duplex connections allows us to have responders started using the on-demand strategy. This assures that once the remote peer start using the connection the local side will be ready to serve it.

For a Duplex connection: a RemoteIdle connection transitions to RemoteCold state after all responders being idle for protocolIdleTimeout. This triggers unregisterInboundConnection.

For a Unidreictional connection: after all responders terminated.

Bundled Patterns

pattern RemoteEstablished ∷ RemoteState m

updateRemoteState ∷ Ord peerAddr ⇒ ConnectionId peerAddr → RemoteState m → InboundGovernorState muxMode peerAddr m a b → InboundGovernorState muxMode peerAddr m a b

Set csRemoteState for a given connection.

mapRemoteState ∷ Ord peerAddr ⇒ ConnectionId peerAddr → (RemoteState m → RemoteState m) → InboundGovernorState muxMode peerAddr m a b → InboundGovernorState muxMode peerAddr m a b

data MiniProtocolData muxMode m a b

Constructors

MiniProtocolData

Fields

mpdMiniProtocol ∷ !(MiniProtocol muxMode ByteString m a b)

Static MiniProtocol description.

mpdMiniProtocolTemp ∷ !ProtocolTemperature

Static mini-protocol temperature.