Ouroboros.Network.Protocol.Handshake

Description

API for running Handshake protocol.

Documentation

runHandshakeClient ∷ (MonadAsync m, MonadFork m, MonadMonotonicTime m, MonadTimer m, MonadMask m, MonadThrow (STM m), Ord vNumber) ⇒ MuxBearer m → connectionId → HandshakeArguments connectionId vNumber vData m → Versions vNumber vData application → m (Either (HandshakeException vNumber) (application, vNumber, vData))

Run client side of the Handshake protocol

runHandshakeServer ∷ (MonadAsync m, MonadFork m, MonadMonotonicTime m, MonadTimer m, MonadMask m, MonadThrow (STM m), Ord vNumber) ⇒ MuxBearer m → connectionId → HandshakeArguments connectionId vNumber vData m → Versions vNumber vData application → m (Either (HandshakeException vNumber) (application, vNumber, vData))

Run server side of the Handshake protocol.

data HandshakeArguments connectionId vNumber vData m

Common arguments for both Handshake client & server.

Constructors

HandshakeArguments

Fields

haHandshakeTracer ∷ Tracer m (WithMuxBearer connectionId (TraceSendRecv (Handshake vNumber Term)))

Handshake tracer

haHandshakeCodec ∷ Codec (Handshake vNumber Term) DeserialiseFailure m ByteString

Codec for protocol messages.

haVersionDataCodec ∷ VersionDataCodec Term vNumber vData

A codec for protocol parameters.

haAcceptVersion ∷ vData → vData → Accept vData

accept version, first argument is our version data the second argument is the remote version data.

haTimeLimits ∷ ProtocolTimeLimits (Handshake vNumber Term)

Driver timeouts for Handshake protocol.

newtype Versions vNum vData r

The version map supported by the local agent keyed on the version identifier.

Each Version contains a function which takes negotiated version data and returns negotiated application (the r type variable).

If one needs to combine multiple versions the simplest way is to use one of the combinators: foldMapVersions, combineVersions or the Semigroup instance directly:

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions ∷ Map vNum (Version vData r)

data HandshakeException vNumber

Wrapper around initiator and responder errors experienced by tryHandshake.

Constructors

HandshakeProtocolLimit ProtocolLimitFailure

HandshakeProtocolError (HandshakeProtocolError vNumber)

data HandshakeProtocolError vNumber

Extends handshake error RefuseReason type, by client specific errors.

Constructors

HandshakeError (RefuseReason vNumber)

NotRecognisedVersion vNumber

InvalidServerSelection vNumber Text

data RefuseReason vNumber

Reasons by which a server can refuse proposed version.

Constructors

VersionMismatch [vNumber] [Int]
All of the prosed versions where not known to the server. Since the server sends all versions that it can knows about, some of them we might not be able to decode, so we include raw tags [Int].

HandshakeDecodeError vNumber Text
The server failed to decode version parameters.

Refused vNumber Text
The server refused to run the proposed version parameters

data Accept vData

A Maybe like type which better explains its purpose.

Constructors

Accept vData

Refuse !Text