Ouroboros.Network.Protocol.Handshake.Codec

Documentation

codecHandshake ∷ ∀ vNumber m failure. (MonadST m, Ord vNumber, Show failure) ⇒ CodecCBORTerm (failure, Maybe Int) vNumber → Codec (Handshake vNumber Term) DeserialiseFailure m ByteString

Handshake codec. The MsgProposeVersions encodes proposed map in ascending order and it expects to receive them in this order. This allows to construct the map in linear time. There is also another limiting factor to the number of versions on can present: the whole message must fit into a single TCP segment.

byteLimitsHandshake ∷ ∀ vNumber. ProtocolSizeLimits (Handshake vNumber Term) ByteString

Byte limits

timeLimitsHandshake ∷ ∀ vNumber. ProtocolTimeLimits (Handshake vNumber Term)

Time limits.

noTimeLimitsHandshake ∷ ∀ vNumber. ProtocolTimeLimits (Handshake vNumber Term)

encodeRefuseReason ∷ CodecCBORTerm fail vNumber → RefuseReason vNumber → Encoding

decodeRefuseReason ∷ Show failure ⇒ CodecCBORTerm (failure, Maybe Int) vNumber → Decoder s (RefuseReason vNumber)

Version data codec

data VersionDataCodec bytes vNumber vData

Codec for version data (vData in code) exchanged by the handshake protocol.

Note: extra type param is instantiated to DictVersion; agreedOptions is instantiated to NodeToNodeVersionData in Ouroboros.Network.NodeToNode or to () in Ouroboros.Network.NodeToClient.

Constructors

VersionDataCodec

Fields

encodeData ∷ vNumber → vData → bytes

encoder of vData which has access to 'extra vData' which can bring extra instances into the scope (by means of pattern matching on a GADT).

decodeData ∷ vNumber → bytes → Either Text vData

decoder of vData.

cborTermVersionDataCodec ∷ (vNumber → CodecCBORTerm Text vData) → VersionDataCodec Term vNumber vData