Ouroboros.Network.Protocol.Handshake.Type

Handshake Protocol

data Handshake vNumber vParams where

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

Constructors

StPropose ∷ Handshake vNumber vParams

StConfirm ∷ Handshake vNumber vParams

StDone ∷ Handshake vNumber vParams

data family Message ps (st ∷ ps) (st' ∷ ps)

The messages for this protocol. It is expected to be a GADT that is indexed by the from and to protocol states. That is the protocol state the message transitions from, and the protocol state it transitions into. These are the edges of the protocol state transition system.

data family ClientHasAgency (st ∷ ps)

Tokens for those protocol states in which the client has agency.

data family ServerHasAgency (st ∷ ps)

Tokens for those protocol states in which the server has agency.

data family NobodyHasAgency (st ∷ ps)

Tokens for terminal protocol states in which neither the client nor server has agency.

On simultaneous open both sides will send MsgProposeVersions, which will be decoded as MsgReplyVersions. It is a terminal message of the protocol. It is important to stress that in this case both sides will make the choice which version and parameters to pick. Our algorithm for picking version is symmetric, which ensures that both sides will end up with the same choice. If one side decides to refuse the version it will close the connection, without sending the reason to the other side.

data RefuseReason vNumber

Reasons by which a server can refuse proposed version.

Constructors

VersionMismatch [vNumber] [Int]
All of the prosed versions where not known to the server. Since the server sends all versions that it can knows about, some of them we might not be able to decode, so we include raw tags [Int].

HandshakeDecodeError vNumber Text
The server failed to decode version parameters.

Refused vNumber Text
The server refused to run the proposed version parameters

data HandshakeProtocolError vNumber

Extends handshake error RefuseReason type, by client specific errors.

Constructors

HandshakeError (RefuseReason vNumber)

NotRecognisedVersion vNumber

InvalidServerSelection vNumber Text