Ouroboros.Network.Protocol.Handshake.Unversioned

Description

Unversioned protocol, used in tests and demo applications.

Documentation

data UnversionedProtocol

Version negotiation for an unversioned protocol. We only use this for tests and demos where proper versioning is excessive.

Constructors

UnversionedProtocol

data UnversionedProtocolData

Constructors

UnversionedProtocolData

unversionedHandshakeCodec ∷ MonadST m ⇒ Codec (Handshake UnversionedProtocol Term) DeserialiseFailure m ByteString

Handshake codec used in various tests.

unversionedProtocolDataCodec ∷ UnversionedProtocol → CodecCBORTerm Text UnversionedProtocolData

unversionedProtocol ∷ app → Versions UnversionedProtocol UnversionedProtocolData app

Make a Versions for an unversioned protocol. Only use this for tests and demos where proper versioning is excessive.