Ouroboros.Network.Protocol.Handshake.Version

Documentation

newtype Versions vNum vData r

The version map supported by the local agent keyed on the version identifier.

Each Version contains a function which takes negotiated version data and returns negotiated application (the r type variable).

If one needs to combine multiple versions the simplest way is to use one of the combinators: foldMapVersions, combineVersions or the Semigroup instance directly:

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions ∷ Map vNum (Version vData r)

data Version vData r

Constructors

Version

Fields

versionApplication ∷ vData → r

versionData ∷ vData

data Accept vData

A Maybe like type which better explains its purpose.

Constructors

Accept vData

Refuse !Text

class Acceptable v where

Methods

acceptableVersion ∷ v → v → Accept v

data VersionMismatch vNum where

Constructors

NoCommonVersion ∷ VersionMismatch vNum

InconsistentVersion ∷ vNum → VersionMismatch vNum

Simple or no versioning

simpleSingletonVersions ∷ vNum → vData → r → Versions vNum vData r

Singleton smart constructor for Versions.

foldMapVersions ∷ (Ord vNum, Foldable f, HasCallStack) ⇒ (x → Versions vNum extra r) → f x → Versions vNum extra r

Useful for folding multiple Versions.

A foldMap restricted to the Versions Semigroup.

PRECONDITION: f x is non-empty.

combineVersions ∷ (Ord vNum, Foldable f, HasCallStack) ⇒ f (Versions vNum extra r) → Versions vNum extra r