Ouroboros.Network.Server.RateLimiting

Description

Rage limiting of accepted connections

Documentation

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

runConnectionRateLimits ∷ (MonadSTM m, MonadDelay m, MonadTime m) ⇒ Tracer m AcceptConnectionsPolicyTrace → STM m Int → AcceptedConnectionsLimit → m ()

Get the number of current connections, make decision based on AcceptedConnectionsLimit and execute it.

Tracing

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32