Ouroboros.Network.Server.Socket

Documentation

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

type BeginConnection addr channel st r = Time → addr → st → STM (HandleConnection channel st r)

What to do on a new connection: accept and run this IO, or reject.

data HandleConnection channel st r where

What to do with a new connection: reject it and give a new state, or accept it and give a new state with a continuation to run against the resulting channel. See also CompleteConnection, which is run for every connection when it finishes, and can also update the state.

Constructors

Reject ∷ !st → HandleConnection channel st r

Accept ∷ !st → !(channel → IO r) → HandleConnection channel st r

type ApplicationStart addr st = addr → Async () → st → STM st

A call back which runs when application starts;

It is needed only because BeginConnection does not have access to the thread which runs the application.

type CompleteConnection addr st tr r = Result addr r → st → STM (CompleteApplicationResult IO addr st)

How to update state when a connection finishes. Can use throwSTM to terminate the server.

TODO: remove async, use `Async m ()` from MonadAsync.

data CompleteApplicationResult m addr s

Constructors

CompleteApplicationResult

Fields

carState ∷ !s

new state

carThreads ∷ Set (Async m ())

threads to kill

carTrace ∷ Maybe (WithAddr addr ErrorPolicyTrace)

trace points

data Result addr r

The product of a spawned thread. We catch all (even async) exceptions.

Constructors

Result

Fields

resultThread ∷ !(Async ())

resultAddr ∷ !addr

resultTime ∷ !Time

resultValue ∷ !(Either SomeException r)

type Main st t = st → STM t

Given a current state, retry unless you want to stop the server. When this transaction returns, any running threads spawned by the server will be killed.

It's possible that a connection is accepted after the main thread returns, but before the server stops. In that case, it will be killed, and the CompleteConnection will not run against it.

run ∷ Tracer IO (WithAddr addr ErrorPolicyTrace) → Tracer IO AcceptConnectionsPolicyTrace → Socket addr channel → AcceptedConnectionsLimit → (IOException → IO ()) → BeginConnection addr channel st r → ApplicationStart addr st → CompleteConnection addr st tr r → Main st t → TVar st → IO t

Run a server.

data Socket addr channel

Abstraction of something that can provide connections. A Socket can be used to get a `Socket SockAddr (Channel IO Lazy.ByteString)` It's not defined in here, though, because we don't want the dependency on typed-protocols or even on network.

Constructors

Socket

Fields

acceptConnection ∷ IO (addr, channel, IO (), Socket addr channel)

The address, a channel, IO to close the channel.

ioSocket ∷ IO (addr, channel) → Socket addr channel

Expected to be useful for testing.