Ouroboros.Network.Server2

Description

Server implementation based on ConnectionManager

Documentation

data ServerArguments (muxMode ∷ MuxMode) socket peerAddr versionNumber bytes m a b

Server static configuration.

Constructors

ServerArguments

Fields

serverSockets ∷ NonEmpty socket

serverSnocket ∷ Snocket m socket peerAddr

serverTracer ∷ Tracer m (ServerTrace peerAddr)

serverTrTracer ∷ Tracer m (RemoteTransitionTrace peerAddr)

serverInboundGovernorTracer ∷ Tracer m (InboundGovernorTrace peerAddr)

serverConnectionLimits ∷ AcceptedConnectionsLimit

serverConnectionManager ∷ MuxConnectionManager muxMode socket peerAddr versionNumber bytes m a b

serverInboundIdleTimeout ∷ DiffTime

Time for which all protocols need to be idle to trigger DemotedToCold transition.

serverControlChannel ∷ ServerControlChannel muxMode peerAddr bytes m a b

Server control var is passed as an argument; this allows to use the server to run and manage responders which needs to be started on inbound connections.

serverObservableStateVar ∷ StrictTVar m InboundGovernorObservableState

Observable mutable state.

newtype InboundGovernorObservableState

Currently only StdGen, but in the future this will be extended to a record which contains some useful statistics about peers to support more advances prune strategies (see. PruneStrategy).

Constructors

InboundGovernorObservableState

Fields

igosPrng ∷ StdGen

newObservableStateVar ∷ MonadLabelledSTM m ⇒ StdGen → m (StrictTVar m InboundGovernorObservableState)

Create new observable state StrictTVar.

newObservableStateVarIO ∷ IO (StrictTVar IO InboundGovernorObservableState)

Using the global StdGen.

newObservableStateVarFromSeed ∷ MonadLabelledSTM m ⇒ Int → m (StrictTVar m InboundGovernorObservableState)

Useful for testing, it is using mkStdGen.

Run server

run ∷ ∀ muxMode socket peerAddr versionNumber m a b. (MonadAsync m, MonadCatch m, MonadEvaluate m, MonadLabelledSTM m, MonadMask m, MonadThrow (STM m), MonadTime m, MonadTimer m, HasResponder muxMode ~ True, Ord peerAddr, Show peerAddr) ⇒ ServerArguments muxMode socket peerAddr versionNumber ByteString m a b → m Void

Run the server, which consists of the following components:

inbound governor, it corresponds to p2p-governor on outbound side

accept loop(s), one per given ip address. We support up to one ipv4 address and up to one ipv6 address, i.e. an ipv6 enabled node will run two accept loops on listening on different addresses with shared inbound governor.

The server can be run in either of two MuxMode-es:

InitiatorResponderMode

ResponderMode

The first one is used in data diffusion for Node-To-Node protocol, while the other is useful for running a server for the Node-To-Client protocol.

Trace

data ServerTrace peerAddr

Constructors

TrAcceptConnection peerAddr

TrAcceptError SomeException

TrAcceptPolicyTrace AcceptConnectionsPolicyTrace

TrServerStarted [peerAddr]

TrServerStopped

TrServerError SomeException
similar to TrAcceptConnection but it is logged once the connection is handed to inbound connection manager, e.g. after handshake negotiation.

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

ServerTraceAcceptConnectionResume Int

data RemoteSt

Remote connection state tracked by inbound protocol governor.

Constructors

RemoteWarmSt

RemoteHotSt

RemoteIdleSt

RemoteColdSt

type RemoteTransition = Transition' (Maybe RemoteSt)

Nothing represents unitialised state.

type RemoteTransitionTrace peerAddr = TransitionTrace' peerAddr (Maybe RemoteSt)

ControlChannel

module Ouroboros.Network.InboundGovernor.ControlChannel