Ouroboros.Network.Socket

Description

Module exports interface for running a node over a socket over TCP / IP.

High level socket interface

data ConnectionTable m addr

data ConnectionTableRef

Constructors

ConnectionTableCreate
No connection to peer exists, attempt to create one.

ConnectionTableExist
A connection to the peer existed, either from another subscriber or the peer opened one towards us.

ConnectionTableDuplicate
This subscriber already has counted a connection to this peer. It must try another target.

data ValencyCounter m

ValencyCounter represents how many active connections we have towards a given peer. It starts out with a positive value representing a desired number of connections for a specific subscription worker. It can become negative, for example if a peer opens multiple connections to us. The vcId is unique per ConnectionTable and ensures that we won't count the same connection twice.

data NetworkMutableState addr

Mutable state maintained by the network component.

Constructors

NetworkMutableState

Fields

nmsConnectionTable ∷ ConnectionTable IO addr

ConnectionTable which maintains information about current upstream and downstream connections.

nmsPeerStates ∷ StrictTVar IO (PeerStates IO addr)

PeerStates which maintains state of each downstream / upstream peer that errored, misbehaved or was not interesting to us.

data SomeResponderApplication bytes m b where

Wrapper for OuroborosResponderApplication and OuroborosInitiatorAndResponderApplication.

Constructors

SomeResponderApplication ∷ ∀ appType m bytes a b. HasResponder appType ~ True ⇒ OuroborosApplication appType bytes m a b → SomeResponderApplication bytes m b

newNetworkMutableState ∷ IO (NetworkMutableState addr)

newNetworkMutableStateSTM ∷ STM (NetworkMutableState addr)

cleanNetworkMutableState ∷ NetworkMutableState addr → IO ()

Clean PeerStates within NetworkMutableState every 200s

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

withServerNode

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber, Ord addr)

⇒ Snocket IO fd addr

→ NetworkServerTracers addr vNumber

→ NetworkMutableState addr

→ AcceptedConnectionsLimit

→ addr

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ (∀ vData. extra vData → vData → vData → Accept)

→ Versions vNumber extra (ConnectionId addr → SomeResponderApplication ByteString IO b)
The mux application that will be run on each incoming connection from a given address. Note that if MuxClientAndServerApplication is returned, the connection will run a full duplex set of mini-protocols.

→ ErrorPolicies

→ (addr → Async Void → IO t)
callback which takes the Async of the thread that is running the server. Note: the server thread will terminate when the callback returns or throws an exception.

→ IO t

Run a server application. It will listen on the given address for incoming connection, otherwise like withServerNode'.

withServerNode'

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber, Ord addr)

⇒ Snocket IO fd addr

→ NetworkServerTracers addr vNumber

→ NetworkMutableState addr

→ AcceptedConnectionsLimit

→ fd

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ (∀ vData. extra vData → vData → vData → Accept)

→ Versions vNumber extra (ConnectionId addr → SomeResponderApplication ByteString IO b)
The mux application that will be run on each incoming connection from a given address. Note that if MuxClientAndServerApplication is returned, the connection will run a full duplex set of mini-protocols.

→ ErrorPolicies

→ (addr → Async Void → IO t)
callback which takes the Async of the thread that is running the server. Note: the server thread will terminate when the callback returns or throws an exception.

→ IO t

Run a server application on the provided socket. The socket must be ready to accept connections. The server thread runs using withAsync function, which means that it will terminate when the callback terminates or throws an exception.

TODO: we should track connections in the state and refuse connections from peers we are already connected to. This is also the right place to ban connection from peers which missbehaved.

The server will run handshake protocol on each incoming connection. We assume that each versin negotiation message should fit into maxTransmissionUnit (~5k bytes).

Note: it will open a socket in the current thread and pass it to the spawned thread which runs the server. This makes it useful for testing, where we need to guarantee that a socket is open before we try to connect to it.

connectToNode

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber, HasInitiator appType ~ True)

⇒ Snocket IO fd addr

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ NetworkConnectTracers addr vNumber

→ Versions vNumber extra (ConnectionId addr → OuroborosApplication appType ByteString IO a b)
application to run over the connection

→ Maybe addr
local address; the created socket will bind to it

→ addr
remote address

→ IO ()

Connect to a remote node. It is using bracket to enclose the underlying socket acquisition. This implies that when the continuation exits the underlying bearer will get closed.

The connection will start with handshake protocol sending Versions to the remote peer. It must fit into maxTransmissionUnit (~5k bytes).

Exceptions thrown by MuxApplication are rethrown by connectTo.

connectToNodeSocket

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber, HasInitiator appType ~ True)

⇒ IOManager

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ NetworkConnectTracers SockAddr vNumber

→ Versions vNumber extra (ConnectionId SockAddr → OuroborosApplication appType ByteString IO a b)
application to run over the connection

→ Socket

→ IO ()

connectToNode'

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber, HasInitiator appType ~ True)

⇒ Snocket IO fd addr

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ NetworkConnectTracers addr vNumber

→ Versions vNumber extra (ConnectionId addr → OuroborosApplication appType ByteString IO a b)
application to run over the connection

→ fd

→ IO ()

Connect to a remote node using an existing socket. It is up to to caller to ensure that the socket is closed in case of an exception.

The connection will start with handshake protocol sending Versions to the remote peer. It must fit into maxTransmissionUnit (~5k bytes).

Exceptions thrown by MuxApplication are rethrown by connectTo.

Traces

data NetworkConnectTracers addr vNumber

Tracer used by connectToNode (and derivatives, like connectTo or 'Ouroboros.Network.NodeToClient.connectTo).

Constructors

NetworkConnectTracers

Fields

nctMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nctHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nullNetworkConnectTracers ∷ NetworkConnectTracers addr vNumber

debuggingNetworkConnectTracers ∷ (Show addr, Show vNumber) ⇒ NetworkConnectTracers addr vNumber

data NetworkServerTracers addr vNumber

Tracers required by a server which handles inbound connections.

Constructors

NetworkServerTracers

Fields

nstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nstAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

tracing rate limiting of accepting connections.

nullNetworkServerTracers ∷ NetworkServerTracers addr vNumber

debuggingNetworkServerTracers ∷ (Show addr, Show vNumber) ⇒ NetworkServerTracers addr vNumber

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

Helper function for creating servers

fromSnocket

Arguments

∷ Ord addr

⇒ ConnectionTable IO addr

→ Snocket IO fd addr

→ fd
socket or handle

→ Socket addr fd

Make a server-compatible socket from a network socket.

beginConnection

Arguments

∷ (Ord vNumber, Typeable vNumber, Show vNumber)

⇒ Snocket IO fd addr

→ Tracer IO (WithMuxBearer peerid MuxTrace)

→ Tracer IO (WithMuxBearer peerid (TraceSendRecv (Handshake vNumber Term)))

→ Codec (Handshake vNumber Term) DeserialiseFailure IO ByteString

→ VersionDataCodec extra Term

→ (∀ vData. extra vData → vData → vData → Accept)

→ (Time → addr → st → STM (AcceptConnection st vNumber extra peerid IO ByteString))
either accept or reject a connection.

→ BeginConnection addr fd st ()

Accept or reject incoming connection based on the current state and address of the incoming connection.

Re-export of PeerStates

data PeerStates m addr

Map from addresses to PeerStates; it will be be shared in a StrictTVar.

Abstracting t is useful for tests, the IO version will use Time IO.

Re-export connection table functions

newConnectionTable ∷ MonadSTM m ⇒ m (ConnectionTable m addr)

refConnection ∷ (MonadSTM m, Ord addr) ⇒ ConnectionTable m addr → addr → ValencyCounter m → m ConnectionTableRef

addConnection

Arguments

∷ (MonadSTM m, Ord addr)

⇒ ConnectionTable m addr

→ addr

→ addr

→ Maybe (ValencyCounter m)
Optional ValencyCounter, used by subscription worker and set to Nothing when called by a local server.

→ STM m ()

Insert a new connection into the ConnectionTable.

removeConnection ∷ ∀ m addr. (MonadSTM m, Ord addr) ⇒ ConnectionTable m addr → addr → addr → m ()

newValencyCounter

Arguments

∷ MonadSTM m

⇒ ConnectionTable m addr

→ Int
Desired valency, that is number of connections a subscription worker will attempt to maintain.

→ STM m (ValencyCounter m)

Create a new ValencyCounter

addValencyCounter ∷ MonadSTM m ⇒ ValencyCounter m → STM m ()

Add a connection.

remValencyCounter ∷ MonadSTM m ⇒ ValencyCounter m → STM m ()

Remove a connection.

waitValencyCounter ∷ MonadSTM m ⇒ ValencyCounter m → STM m ()

Wait until ValencyCounter becomes positive, used for detecting when we can create new connections.

readValencyCounter ∷ MonadSTM m ⇒ ValencyCounter m → STM m Int

Returns current ValencyCounter value, represent the number of additional connections that can be created. May be negative.

Auxiliary functions

sockAddrFamily ∷ SockAddr → Family