Ouroboros.Network.Subscription

Description

Public interface of Subscription workers.

IP Subscription Worker

ipSubscriptionWorker ∷ ∀ a. Snocket IO Socket SockAddr → Tracer IO (WithIPList (SubscriptionTrace SockAddr)) → Tracer IO (WithAddr SockAddr ErrorPolicyTrace) → NetworkMutableState SockAddr → IPSubscriptionParams a → (Socket → IO a) → IO Void

Spawns a subscription worker which will attempt to keep the specified number of connections (Valency) active towards the list of IP addresses given in IPSubscriptionTarget.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

DNS Subscription Worker

dnsSubscriptionWorker ∷ Snocket IO Socket SockAddr → Tracer IO (WithDomainName (SubscriptionTrace SockAddr)) → Tracer IO (WithDomainName DnsTrace) → Tracer IO (WithAddr SockAddr ErrorPolicyTrace) → NetworkMutableState SockAddr → DnsSubscriptionParams a → (Socket → IO a) → IO Void

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data ConnectResult

GADT which classifies connection result.

Constructors

ConnectSuccess
Successful connection.

ConnectSuccessLast
Successfully connection, reached the valency target. Other ongoing connection attempts will be killed.

ConnectValencyExceeded
Someone else manged to create the final connection to a target before us.

Constants

defaultConnectionAttemptDelay ∷ DiffTime

Time to wait between connection attempts when we don't have any DeltaQ info.

minConnectionAttemptDelay ∷ DiffTime

Minimum time to wait between connection attempts.

maxConnectionAttemptDelay ∷ DiffTime

Maximum time to wait between connection attempts.

ipRetryDelay ∷ DiffTime

Minimum time to wait between ip reconnects

resolutionDelay ∷ DiffTime

Time to wait for an AAAA response after receiving an A response.

Errors

data SubscriberError

Constructors

SubscriberError

Fields

seType ∷ !SubscriberErrorType

seMessage ∷ !String

seStack ∷ !CallStack

Tracing

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e ⇒ SubscriptionTraceSocketAllocationException addr e

Exception e ⇒ SubscriptionTraceConnectException addr e

Exception e ⇒ SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

data WithIPList a

Constructors

WithIPList

Fields

wilSrc ∷ !(LocalAddresses SockAddr)

wilDsts ∷ ![SockAddr]

wilEvent ∷ !a

data DnsTrace

Constructors

DnsTraceLookupException SomeException

DnsTraceLookupAError DNSError

DnsTraceLookupAAAAError DNSError

DnsTraceLookupIPv6First

DnsTraceLookupIPv4First

DnsTraceLookupAResult [SockAddr]

DnsTraceLookupAAAAResult [SockAddr]

data WithDomainName a

Constructors

WithDomainName

Fields

wdnDomain ∷ !Domain

wdnEvent ∷ !a