Ouroboros.Network.Subscription.Dns

Documentation

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data Resolver m

Constructors

Resolver

Fields

lookupA ∷ Domain → m (Either DNSError [SockAddr])

lookupAAAA ∷ Domain → m (Either DNSError [SockAddr])

type DnsSubscriptionParams a = SubscriptionParams a DnsSubscriptionTarget

dnsSubscriptionWorker' ∷ Snocket IO Socket SockAddr → Tracer IO (WithDomainName (SubscriptionTrace SockAddr)) → Tracer IO (WithDomainName DnsTrace) → Tracer IO (WithAddr SockAddr ErrorPolicyTrace) → NetworkMutableState SockAddr → Resolver IO → DnsSubscriptionParams a → Main IO (PeerStates IO SockAddr) x → (Socket → IO a) → IO x

dnsSubscriptionWorker ∷ Snocket IO Socket SockAddr → Tracer IO (WithDomainName (SubscriptionTrace SockAddr)) → Tracer IO (WithDomainName DnsTrace) → Tracer IO (WithAddr SockAddr ErrorPolicyTrace) → NetworkMutableState SockAddr → DnsSubscriptionParams a → (Socket → IO a) → IO Void

dnsResolve ∷ ∀ m s. (MonadAsync m, MonadSay m, MonadSTM m, MonadTime m, MonadTimer m, MonadThrow m) ⇒ Tracer m DnsTrace → Resolver m → StrictTVar m s → BeforeConnect m s SockAddr → DnsSubscriptionTarget → m (SubscriptionTarget m SockAddr)

resolutionDelay ∷ DiffTime

Time to wait for an AAAA response after receiving an A response.

Traces

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e ⇒ SubscriptionTraceSocketAllocationException addr e

Exception e ⇒ SubscriptionTraceConnectException addr e

Exception e ⇒ SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

data DnsTrace

Constructors

DnsTraceLookupException SomeException

DnsTraceLookupAError DNSError

DnsTraceLookupAAAAError DNSError

DnsTraceLookupIPv6First

DnsTraceLookupIPv4First

DnsTraceLookupAResult [SockAddr]

DnsTraceLookupAAAAResult [SockAddr]

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime !DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithDomainName a

Constructors

WithDomainName

Fields

wdnDomain ∷ !Domain

wdnEvent ∷ !a

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr ∷ !addr

wiaEvent ∷ !a