Ouroboros.Network.Subscription.Ip

Description

IP subscription worker implentation.

Documentation

data SubscriptionParams a target

ipSubscriptionWorker and dnsSubscriptionWorker parameters

Constructors

SubscriptionParams

Fields

spLocalAddresses ∷ LocalAddresses SockAddr

spConnectionAttemptDelay ∷ SockAddr → Maybe DiffTime

should return expected delay for the given address

spErrorPolicies ∷ ErrorPolicies

spSubscriptionTarget ∷ target

type IPSubscriptionParams a = SubscriptionParams a IPSubscriptionTarget

ipSubscriptionWorker ∷ ∀ a. Snocket IO Socket SockAddr → Tracer IO (WithIPList (SubscriptionTrace SockAddr)) → Tracer IO (WithAddr SockAddr ErrorPolicyTrace) → NetworkMutableState SockAddr → IPSubscriptionParams a → (Socket → IO a) → IO Void

Spawns a subscription worker which will attempt to keep the specified number of connections (Valency) active towards the list of IP addresses given in IPSubscriptionTarget.

subscriptionWorker

Arguments

∷ Snocket IO Socket SockAddr

→ Tracer IO (SubscriptionTrace SockAddr)

→ Tracer IO (WithAddr SockAddr ErrorPolicyTrace)

→ NetworkMutableState SockAddr

→ WorkerParams IO LocalAddresses SockAddr

→ ErrorPolicies

→ Main IO (PeerStates IO SockAddr) x
main callback

→ (Socket → IO a)
application to run on each connection

→ IO x

Like worker but in IO; It provides address selection function, SocketStateChange and CompleteApplication callbacks. The Main callback is left as it's useful for testing purposes.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

ipSubscriptionTarget ∷ ∀ m addr. (MonadSTM m, MonadTime m, Ord addr) ⇒ Tracer m (SubscriptionTrace addr) → StrictTVar m (PeerStates m addr) → [addr] → SubscriptionTarget m addr

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e ⇒ SubscriptionTraceSocketAllocationException addr e

Exception e ⇒ SubscriptionTraceConnectException addr e

Exception e ⇒ SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime !DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithIPList a

Constructors

WithIPList

Fields

wilSrc ∷ !(LocalAddresses SockAddr)

wilDsts ∷ ![SockAddr]

wilEvent ∷ !a

PeerState STM transactions

type BeforeConnect m s addr = Time → addr → s → STM m (ConnectDecision s)

Check state before connecting to a remote peer. We will connect only if it retuns True.

runBeforeConnect ∷ (MonadSTM m, MonadTime m) ⇒ StrictTVar m s → BeforeConnect m s addr → addr → m Bool

Run BeforeConnect callback in a MonadTime monad.

beforeConnectTx ∷ ∀ m addr. (MonadSTM m, Ord addr) ⇒ BeforeConnect m (PeerStates m addr) addr

BeforeConnect callback: it updates peer state and return boolean value wheather to connect to it or not. If a peer hasn't been recorded in PeerStates, we add it and try to connect to it.

completeApplicationTx ∷ ∀ m addr a. (MonadAsync m, Ord addr, Ord (Async m ())) ⇒ ErrorPolicies → CompleteApplication m (PeerStates m addr) addr a

CompleteApplication callback

socketStateChangeTx ∷ (Ord addr, Show addr) ⇒ SocketStateChange IO (PeerStates IO addr) addr

mainTx ∷ (MonadThrow m, MonadThrow (STM m), MonadSTM m) ⇒ Main m (PeerStates m addr) Void

Main callback. It throws an exception when the state becomes ThrowException. This exception is thrown from the main thread.

Utilitity functions

selectSockAddr ∷ SockAddr → LocalAddresses SockAddr → Maybe SockAddr