Ouroboros.Network.Subscription.Subscriber

Documentation

newtype SubscriptionTarget m target

Generate subscription targets in some monad. Examples include obtaining targets from a fixed list, or from a DNS lookup.

Constructors

SubscriptionTarget

Fields

getSubscriptionTarget ∷ m (Maybe (target, SubscriptionTarget m target))

This should be used with the exception that implementations can block on the order of seconds.

listSubscriptionTarget ∷ Applicative m ⇒ [target] → SubscriptionTarget m target