Ouroboros.Network.Subscription.Worker

Documentation

type SocketStateChange m s addr = SocketState m addr → s → STM m s

Callback which fires: when we create or close a socket.

data SocketState m addr

Constructors

CreatedSocket !addr !(Async m ())

ClosedSocket !addr !(Async m ())

type CompleteApplication m s addr r = Result addr r → s → STM m (CompleteApplicationResult m addr s)

Complete a connection, which receive application result (or exception).

data ConnectResult

GADT which classifies connection result.

Constructors

ConnectSuccess
Successful connection.

ConnectSuccessLast
Successfully connection, reached the valency target. Other ongoing connection attempts will be killed.

ConnectValencyExceeded
Someone else manged to create the final connection to a target before us.

data Result addr r where

Result of the connection thread. It's either result of an application, or an exception thrown by it.

Constructors

ApplicationResult ∷ !Time → !addr → !r → Result addr r

Connected ∷ !Time → !addr → Result addr r

ConnectionError ∷ Exception e ⇒ !Time → !addr → !e → Result addr r

ApplicationError ∷ Exception e ⇒ !Time → !addr → !e → Result addr r

type Main m s t = s → STM m t

Given current state retry too keep the subscription worker going. When this transaction returns, all the threads spawned by the worker will be killed.

type StateVar m s = StrictTVar m s

Mutable state kept by the worker. All the workers in this module are polymorphic over the state type. The state is updated with two callbacks:

CompleteConnect - STM transaction which runs when the connect call returned, if it thrown an exception it will be passed to the callback.

CompleteApplication - STM transaction which runs when application returned. It will receive the result of the application or an exception raised by it.

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 ∷ Maybe addr

Local IPv4 address to use, Nothing indicates don't use IPv4

laIpv6 ∷ Maybe addr

Local IPv6 address to use, Nothing indicates don't use IPv6

laUnix ∷ Maybe addr

Local Unix address to use, Nothing indicates don't use Unix sockets

Subscription worker

data WorkerCallbacks m s addr a t

Worker STM callbacks

Constructors

WorkerCallbacks

Fields

wcSocketStateChangeTx ∷ SocketStateChange m s addr

wcCompleteApplicationTx ∷ CompleteApplication m s addr a

wcMainTx ∷ Main m s t

data WorkerParams m localAddrs addr

Worker parameters

Constructors

WorkerParams

Fields

wpLocalAddresses ∷ localAddrs addr

local addresses of the server

wpSelectAddress ∷ addr → localAddrs addr → Maybe addr

given remote addr pick the local address

wpConnectionAttemptDelay ∷ addr → Maybe DiffTime

delay after a connection attempt to addr

wpSubscriptionTarget ∷ m (SubscriptionTarget m addr)

wpValency ∷ Int

worker

Arguments

∷ (Ord addr, Show addr)

⇒ Tracer IO (SubscriptionTrace addr)

→ Tracer IO (WithAddr addr ErrorPolicyTrace)

→ ConnectionTable IO addr

→ StateVar IO s

→ Snocket IO sock addr

→ WorkerCallbacks IO s addr a x

→ WorkerParams IO localAddrs addr

→ (sock → IO a)
application

→ IO x

This is the most abstract worker, which puts all the pieces together. It will execute until main :: Main m s t returns. It runs subscriptionLoop in a new threads and will exit when it dies. Spawn threads are cancelled in a finally callback by throwing SubscriberError.

Note: This function runs in IO only because MonadSTM does not yet support orElse, PR #432.

Socket API

safeConnect

Arguments

∷ (MonadThrow m, MonadMask m)

⇒ Snocket m sock addr

→ addr
remote addr

→ addr
local addr

→ m ()
allocate extra action; executed with async exceptions masked in the allocation action of bracket

→ m ()
release extra action; executed with async exceptions masked in the closing action of bracket

→ ((∀ x. m x → m x) → sock → Either SomeException () → m t)
continuation executed with async exceptions masked; it receives: unmask function, allocated socket and connection error.

→ m t

Allocate a socket and connect to a peer, execute the continuation with async exceptions masked. The continuation receives the unmask callback.

Constants

defaultConnectionAttemptDelay ∷ DiffTime

Time to wait between connection attempts when we don't have any DeltaQ info.

minConnectionAttemptDelay ∷ DiffTime

Minimum time to wait between connection attempts.

maxConnectionAttemptDelay ∷ DiffTime

Maximum time to wait between connection attempts.

ipRetryDelay ∷ DiffTime

Minimum time to wait between ip reconnects

Errors

data SubscriberError

Constructors

SubscriberError

Fields

seType ∷ !SubscriberErrorType

seMessage ∷ !String

seStack ∷ !CallStack

Tracing

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e ⇒ SubscriptionTraceSocketAllocationException addr e

Exception e ⇒ SubscriptionTraceConnectException addr e

Exception e ⇒ SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr