Simulation.Network.Snocket

Description

This module provides simulation environment and a snocket implementation suitable for IOSim.

Though this module is designed for simulation / testing, it lives in the library, since it is needed in `ouroboros-network-framework:test` and `ouroboros-network:test' components.

TODO: Create a snocket package, in order to avoid having to have ouroboros-network-testing as a dependency for this cabal library.

Simulated Snocket

withSnocket ∷ ∀ m peerAddr a. (MonadCatch m, MonadLabelledSTM m, MonadMask m, MonadTime m, MonadTimer m, MonadThrow (STM m), GlobalAddressScheme peerAddr, Ord peerAddr, Typeable peerAddr, Show peerAddr) ⇒ Tracer m (WithAddr (TestAddress peerAddr) (SnocketTrace m (TestAddress peerAddr))) → BearerInfo → Map (NormalisedId (TestAddress peerAddr)) (Script BearerInfo) → (Snocket m (FD m (TestAddress peerAddr)) (TestAddress peerAddr) → m (ObservableNetworkState (TestAddress peerAddr)) → m a) → m a

A bracket which runs a network simulation. When the simulation terminates it verifies that all listening sockets and all connections are closed. It might throw ResourceException.

newtype ObservableNetworkState addr

Simulation accessible network environment consumed by simSnocket.

Constructors

ObservableNetworkState

Fields

onsConnections ∷ Map (NormalisedId addr) addr

Registry of active connections and respective provider

data ResourceException addr

Constructors

NotReleasedListeningSockets [addr] (Maybe SomeException)

NotReleasedConnections (Map (NormalisedId addr) ConnectionState) (Maybe SomeException)

data SDUSize

newtype Script a

Constructors

Script (NonEmpty a)

type Size = Int64

data SnocketTrace m addr

Constructors

STConnecting (FD_ m addr) addr

STConnected (FD_ m addr) OpenType

STBearerInfo BearerInfo

STConnectError (FD_ m addr) addr IOError

STConnectTimeout TimeoutDetail

STBindError (FD_ m addr) addr IOError

STClosing SockType (Wedge (ConnectionId addr) [addr])

STClosed SockType (Maybe (Maybe ConnectionState))
TODO: Document meaning of 'Maybe (Maybe OpenState)'

STClosingQueue Bool

STClosedQueue Bool

STClosedWhenReading

STAcceptFailure SockType SomeException

STAccepted addr

STBearer (FD_ m addr)

STAttenuatedChannelTrace (ConnectionId addr) AttenuatedChannelTrace

data SockType

Constructors

ListeningSock

ConnectionSock

UnknownType

data OpenType

Either simultaneous open or normal open. Unlike in TCP, only one side will will know that it is doing simultaneous open.

Constructors

SimOpen
Simultaneous open

NormalOpen
Normal open

normaliseId ∷ Ord addr ⇒ ConnectionId addr → NormalisedId addr

Safe constructor of NormalisedId

data BearerInfo

Each bearer info describes outbound and inbound side of a point to point bearer.

Constructors

BearerInfo

Fields

biConnectionDelay ∷ !DiffTime

How long it take to create a connection

biInboundAttenuation ∷ Time → Size → (DiffTime, SuccessOrFailure)

attenuation of inbound side of the bearer, i.e. attenuation used by bearers that were accepted.

biOutboundAttenuation ∷ Time → Size → (DiffTime, SuccessOrFailure)

attenuation of outbound side of the bearer, i.e. the attenuation used by bearers that were created with connect call.

biInboundWriteFailure ∷ !(Maybe Int)

Maximum number of successful writes for an inbound bearer.

biOutboundWriteFailure ∷ !(Maybe Int)

Maximum number of successful writes for an outbound bearer.

biAcceptFailures ∷ !(Maybe (DiffTime, IOErrType, IOErrThrowOrReturn))

Time after which accept will throw an exception.

Currently it only supports two kinds of exceptions, ones that are caught and rethrown by the server (ECONNABORTED), and an exception which would be caught, and delivered to the application via AcceptFailure.

biSDUSize ∷ !SDUSize

SDU size of the bearer; it will be shared between outbound and inbound sides.

Note: shrinking SDUSize means make it larger, as this allows to send fewer chunks through the bearer.

data IOErrType

Error types.

Constructors

IOErrConnectionAborted

IOErrResourceExhausted

data IOErrThrowOrReturn

The io error can be either thrown by accept or return as part of AcceptFailure. Even if our Snocket implementation is faulty, we can verify that the server will behave as expected.

Constructors

IOErrThrow

IOErrReturn

data SuccessOrFailure

Constructors

Success

Failure

data TimeoutDetail

Constructors

WaitingToConnect

WaitingToBeAccepted

noAttenuation ∷ BearerInfo

BearerInfo without attenuation and instantaneous connect delay. It also using the production value of SDUSize.

data FD m peerAddr

File descriptor type.

class GlobalAddressScheme addr where

A type class for global IP address scheme. Every node in the simulation has an ephemeral address. Every node in the simulation has an implicity ipv4 and ipv6 address (if one is not bound by explicitly).

Methods

getAddressType ∷ TestAddress addr → AddressType

ephemeralAddress ∷ AddressType → Natural → TestAddress addr

data AddressType

Connection manager supports IPv4 and IPv6 addresses.

Constructors

IPv4Address

IPv6Address