Ouroboros.Network.Driver

Description

Drivers for running Peers with a Codec and a Channel.

Normal peers

runPeer ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Channel m bytes → Peer ps pr st m a → m a

Run a peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

runPeerWithLimits ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadAsync m, MonadFork m, MonadMask m, MonadThrow (STM m), MonadTime m, MonadTimer m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → ProtocolSizeLimits ps bytes → ProtocolTimeLimits ps → Channel m bytes → Peer ps pr st m a → m a

data TraceSendRecv ps where

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg ∷ AnyMessage ps → TraceSendRecv ps

TraceRecvMsg ∷ AnyMessage ps → TraceSendRecv ps

Pipelined peers

runPipelinedPeer ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadSTM m, MonadAsync m, MonadThrow m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Channel m bytes → PeerPipelined ps pr st m a → m a

Run a pipelined peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

Unlike normal peers, running pipelined peers rely on concurrency, hence the MonadSTM constraint.

runPipelinedPeerWithLimits ∷ ∀ ps (st ∷ ps) pr failure bytes m a. (MonadAsync m, MonadFork m, MonadMask m, MonadThrow (STM m), MonadTime m, MonadTimer m, Exception failure) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → ProtocolSizeLimits ps bytes → ProtocolTimeLimits ps → Channel m bytes → PeerPipelined ps pr st m a → m a

Run a pipelined peer with the given channel via the given codec.

This runs the peer to completion (if the protocol allows for termination).

Unlike normal peers, running pipelined peers rely on concurrency, hence the MonadSTM constraint.