Ouroboros.Network.InboundGovernor.ControlChannel

Description

Intended to be imported qualified.

Documentation

data NewConnection peerAddr handle

Announcment message for a new connection.

Constructors

NewConnection !Provenance !(ConnectionId peerAddr) !DataFlow !handle
Announce a new connection. Inbound protocol governor will start responder protocols using StartOnDemand strategy and monitor remote transitions: PromotedToWarm^{Duplex}_{Remote} and DemotedToCold^{dataFlow}_{Remote}.

data ControlChannel m msg

Server control channel. It allows to pass STM transactions which will resolve to NewConnection. Server's monitoring thread is the consumer of this messages; there are two produceres: accept loop and connection handler for outbound connections.

Constructors

ControlChannel

Fields

readMessage ∷ STM m msg

Read a single NewConnection instructrion from the channel.

writeMessage ∷ msg → STM m ()

Write a NewConnection to the channel.

type ServerControlChannel (muxMode ∷ MuxMode) peerAddr bytes m a b = ControlChannel m (NewConnection peerAddr (Handle muxMode peerAddr bytes m a b))

newControlChannel ∷ ∀ m srvCntrlMsg. MonadLabelledSTM m ⇒ m (ControlChannel m srvCntrlMsg)

newOutboundConnection ∷ ControlChannel m (NewConnection peerAddr handle) → ConnectionId peerAddr → DataFlow → handle → STM m ()

newInboundConnection ∷ ControlChannel m (NewConnection peerAddr handle) → ConnectionId peerAddr → DataFlow → handle → STM m ()