Ouroboros.Network.Mux

Documentation

data MuxMode where

Constructors

InitiatorMode ∷ MuxMode

ResponderMode ∷ MuxMode

InitiatorResponderMode ∷ MuxMode

newtype OuroborosApplication (mode ∷ MuxMode) addr bytes m a b

Like MuxApplication but using a MuxPeer rather than a raw Channel -> m a action.

Constructors

OuroborosApplication (ConnectionId addr → ControlMessageSTM m → [MiniProtocol mode bytes m a b])

type MuxProtocolBundle (mode ∷ MuxMode) addr bytes m a b = ConnectionId addr → ControlMessageSTM m → [MiniProtocol mode bytes m a b]

data ProtocolTemperature

There are three kinds of applications: warm, hot and established (ones that run in for both warm and hot peers).

Constructors

Established

Warm

Hot

data TokProtocolTemperature (pt ∷ ProtocolTemperature) where

Singletons for AppKind

Constructors

TokHot ∷ TokProtocolTemperature Hot

TokWarm ∷ TokProtocolTemperature Warm

TokEstablished ∷ TokProtocolTemperature Established

data SomeTokProtocolTemperature where

Constructors

SomeTokProtocolTemperature ∷ TokProtocolTemperature pt → SomeTokProtocolTemperature

data WithProtocolTemperature (pt ∷ ProtocolTemperature) a where

We keep hot, warm and established application (or their context) distinct. It's only needed for a handly projectBundle map.

Constructors

WithHot ∷ !a → WithProtocolTemperature Hot a

WithWarm ∷ !a → WithProtocolTemperature Warm a

WithEstablished ∷ !a → WithProtocolTemperature Established a

withoutProtocolTemperature ∷ WithProtocolTemperature pt a → a

data WithSomeProtocolTemperature a where

Constructors

WithSomeProtocolTemperature ∷ WithProtocolTemperature pt a → WithSomeProtocolTemperature a

withoutSomeProtocolTemperature ∷ WithSomeProtocolTemperature a → a

data Bundle a

A bundle of HotApp, WarmApp and EstablishedApp.

Constructors

Bundle

Fields

withHot ∷ !(WithProtocolTemperature Hot a)

hot mini-protocols

withWarm ∷ !(WithProtocolTemperature Warm a)

warm mini-protocols

withEstablished ∷ !(WithProtocolTemperature Established a)

established mini-protocols

projectBundle ∷ TokProtocolTemperature pt → Bundle a → a

type OuroborosBundle (mode ∷ MuxMode) addr bytes m a b = Bundle (MuxProtocolBundle mode addr bytes m a b)

type MuxBundle (mode ∷ MuxMode) bytes m a b = Bundle [MiniProtocol mode bytes m a b]

data MiniProtocol (mode ∷ MuxMode) bytes m a b

Constructors

MiniProtocol

Fields

miniProtocolNum ∷ !MiniProtocolNum

miniProtocolLimits ∷ !MiniProtocolLimits

miniProtocolRun ∷ !(RunMiniProtocol mode bytes m a b)

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue ∷ !Int

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

data RunMiniProtocol (mode ∷ MuxMode) bytes m a b where

Constructors

InitiatorProtocolOnly ∷ MuxPeer bytes m a → RunMiniProtocol InitiatorMode bytes m a Void

ResponderProtocolOnly ∷ MuxPeer bytes m b → RunMiniProtocol ResponderMode bytes m Void b

InitiatorAndResponderProtocol ∷ MuxPeer bytes m a → MuxPeer bytes m b → RunMiniProtocol InitiatorResponderMode bytes m a b

data MuxPeer bytes m a where

Constructors

MuxPeer ∷ ∀ (pr ∷ PeerRole) ps (st ∷ ps) failure bytes m a. (Show failure, ∀ (st' ∷ ps). Show (ClientHasAgency st'), ∀ (st' ∷ ps). Show (ServerHasAgency st'), ShowProxy ps) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Peer ps pr st m a → MuxPeer bytes m a

MuxPeerPipelined ∷ ∀ (pr ∷ PeerRole) ps (st ∷ ps) failure bytes m a. (Show failure, ∀ (st' ∷ ps). Show (ClientHasAgency st'), ∀ (st' ∷ ps). Show (ServerHasAgency st'), ShowProxy ps) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → PeerPipelined ps pr st m a → MuxPeer bytes m a

MuxPeerRaw ∷ (Channel m bytes → m (a, Maybe bytes)) → MuxPeer bytes m a

runMuxPeer ∷ (MonadCatch m, MonadAsync m) ⇒ MuxPeer bytes m a → Channel m bytes → m (a, Maybe bytes)

Run a MuxPeer using either runPeer or runPipelinedPeer.

toApplication ∷ (MonadCatch m, MonadAsync m) ⇒ ConnectionId addr → ControlMessageSTM m → OuroborosApplication mode addr ByteString m a b → MuxApplication mode m a b

mkMuxApplicationBundle ∷ ∀ mode addr bytes m a b. ConnectionId addr → Bundle (ControlMessageSTM m) → OuroborosBundle mode addr bytes m a b → MuxBundle mode bytes m a b

mkMiniProtocolBundle ∷ MuxBundle mode bytes m a b → MiniProtocolBundle mode

Make MiniProtocolBundle, which is used to create a mux interface with newMux. The output of mkMuxApplicationBundle can be used as input.

data ControlMessage

Control signal sent to a mini-protocol. expected to exit, on Continue it should continue its operation

Constructors

Continue
Continue operation.

Quiesce
Hold on, e.g. do not sent messages until resumed. This is not used for any hot protocol.

Terminate
The client is expected to terminate as soon as possible.

type ControlMessageSTM m = STM m ControlMessage

ControlMessageSTM should depend on muxMode (we only need to shedule stop for intiator side). This is not done only because this would break tests, but once the old api is removed it should be possible.

continueForever ∷ Applicative (STM m) ⇒ proxy m → ControlMessageSTM m

timeoutWithControlMessage ∷ MonadSTM m ⇒ ControlMessageSTM m → STM m a → m (Maybe a)

First to finish synchronisation between Terminate state of ControlMessage and an stm action.

This should return STM m (Maybe a) but STM is a non-injective type family, and we would need to pass Proxy m to fix an ambiuous type (or use AllowAmbiguousTypes extension).

Re-exports

from Network.Mux

data MuxError

Error type used in accross the mux layer.

Constructors

MuxError

Fields

errorType ∷ !MuxErrorType

errorMsg ∷ !String

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

MuxSDUWriteTimeout
thrown when writing a single SDU takes too long

MuxShutdown !(Maybe MuxErrorType)
Result of runMiniProtocol's completionAction in case of an error or mux being closed while a mini-protocol was still running, this is not a clean exit.

MuxCleanShutdown
Mux stopped by stopMux

type family HasInitiator (mode ∷ MuxMode) ∷ Bool where ...

Equations

HasInitiator 'InitiatorMode = 'True

HasInitiator 'ResponderMode = 'False

HasInitiator 'InitiatorResponderMode = 'True

type family HasResponder (mode ∷ MuxMode) ∷ Bool where ...

Equations

HasResponder 'InitiatorMode = 'False

HasResponder 'ResponderMode = 'True

HasResponder 'InitiatorResponderMode = 'True