Ouroboros.Network.Mux

Documentation

data AppType

Constructors

InitiatorApp

ResponderApp

InitiatorAndResponderApp

newtype OuroborosApplication (appType ∷ AppType) bytes m a b

Like MuxApplication but using a MuxPeer rather than a raw Channel -> m a action.

Constructors

OuroborosApplication [MiniProtocol appType bytes m a b]

data MiniProtocol (appType ∷ AppType) bytes m a b

Constructors

MiniProtocol

Fields

miniProtocolNum ∷ !MiniProtocolNum

miniProtocolLimits ∷ !MiniProtocolLimits

miniProtocolRun ∷ !(RunMiniProtocol appType bytes m a b)

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue ∷ !Int64

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

data RunMiniProtocol (appType ∷ AppType) bytes m a b where

Constructors

InitiatorProtocolOnly ∷ MuxPeer bytes m a → RunMiniProtocol InitiatorApp bytes m a Void

ResponderProtocolOnly ∷ MuxPeer bytes m b → RunMiniProtocol ResponderApp bytes m Void b

InitiatorAndResponderProtocol ∷ MuxPeer bytes m a → MuxPeer bytes m b → RunMiniProtocol InitiatorAndResponderApp bytes m a b

data MuxPeer bytes m a where

Constructors

MuxPeer ∷ Exception failure ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Peer ps pr st m a → MuxPeer bytes m a

MuxPeerPipelined ∷ Exception failure ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → PeerPipelined ps pr st m a → MuxPeer bytes m a

MuxPeerRaw ∷ (Channel m bytes → m a) → MuxPeer bytes m a

toApplication ∷ (MonadCatch m, MonadAsync m) ⇒ OuroborosApplication appType ByteString m a b → MuxApplication appType m a b

Re-exports

from Network.Mux

data MuxError

Error type used in accross the mux layer.

Constructors

MuxError

Fields

errorType ∷ !MuxErrorType

errorMsg ∷ !String

errorStack ∷ !CallStack

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

type family HasInitiator (appType ∷ AppType) ∷ Bool where ...

Equations

HasInitiator InitiatorApp = True

HasInitiator ResponderApp = False

HasInitiator InitiatorAndResponderApp = True

type family HasResponder (appType ∷ AppType) ∷ Bool where ...

Equations

HasResponder InitiatorApp = False

HasResponder ResponderApp = True

HasResponder InitiatorAndResponderApp = True