Ouroboros.Network.Protocol.Handshake.Codec

Documentation

codecHandshake ∷ ∀ vNumber m. (Monad m, MonadST m, Ord vNumber, Enum vNumber, Serialise vNumber, Show vNumber) ⇒ Codec (Handshake vNumber Term) DeserialiseFailure m ByteString

Handshake codec. The MsgProposeVersions encodes proposed map in ascending order and it expects to receive them in this order. This allows to construct the map in linear time. There is also another limiting factor to the number of versions on can present: the whole message must fit into a single TCP segment.

byteLimitsHandshake ∷ ProtocolSizeLimits (Handshake vNumber Term) ByteString

Byte limits

timeLimitsHandshake ∷ ProtocolTimeLimits (Handshake vNumber Term)