Ouroboros.Network.Protocol.Handshake.Codec

Documentation

codecHandshake ∷ ∀ vNumber m failure. (Monad m, MonadThrow m, MonadST m, Ord vNumber, Enum vNumber, Show vNumber, Show failure) ⇒ CodecCBORTerm (failure, Maybe Int) vNumber → Codec (Handshake vNumber Term) DeserialiseFailure m ByteString

Handshake codec. The MsgProposeVersions encodes proposed map in ascending order and it expects to receive them in this order. This allows to construct the map in linear time. There is also another limiting factor to the number of versions on can present: the whole message must fit into a single TCP segment.

byteLimitsHandshake ∷ ProtocolSizeLimits (Handshake vNumber Term) ByteString

Byte limits

timeLimitsHandshake ∷ ProtocolTimeLimits (Handshake vNumber Term)

encodeRefuseReason ∷ CodecCBORTerm fail vNumber → RefuseReason vNumber → Encoding

decodeRefuseReason ∷ Show failure ⇒ CodecCBORTerm (failure, Maybe Int) vNumber → Decoder s (RefuseReason vNumber)

Version data codec

data VersionDataCodec extra bytes

Codec for version data (vData in code) exchanged by the handshake protocol.

Constructors

VersionDataCodec

Fields

encodeData ∷ ∀ vData. extra vData → vData → bytes

encoder of vData which has access to 'extra vData' which can bring extra instances into the scope (by means of pattern matching on a GADT).

decodeData ∷ ∀ vData. extra vData → bytes → Either Text vData

decoder of vData.

cborTermVersionDataCodec ∷ VersionDataCodec DictVersion Term