Ouroboros.Network.Protocol.Handshake.Type

Handshake Protocol

data Handshake vNumber vParams where

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

Constructors

StPropose ∷ Handshake vNumber vParams

StConfirm ∷ Handshake vNumber vParams

StDone ∷ Handshake vNumber vParams

data family Message ps (st ∷ ps) (st' ∷ ps) ∷ Type

The messages for this protocol. It is expected to be a GADT that is indexed by the from and to protocol states. That is the protocol state the message transitions from, and the protocol state it transitions into. These are the edges of the protocol state transition system.

data family ClientHasAgency (st ∷ ps) ∷ Type

Tokens for those protocol states in which the client has agency.

data family ServerHasAgency (st ∷ ps) ∷ Type

Tokens for those protocol states in which the server has agency.

data family NobodyHasAgency (st ∷ ps) ∷ Type

Tokens for terminal protocol states in which neither the client nor server has agency.

data RefuseReason vNumber

Reasons by which a server can refuse proposed version.

Constructors

VersionMismatch [vNumber]
All of the prosed versions where not known to the server.

HandshakeDecodeError vNumber Text
The server failed to decode version parameters.

Refused vNumber Text
The server refused to run the proposed version parameters

data HandshakeClientProtocolError vNumber

Client errors, which extends handshake error RefuseReason type, by client specific errors.

Constructors

HandshakeError (RefuseReason vNumber)

NotRecognisedVersion vNumber

Handshake client

handshakeClientPeer ∷ Ord vNumber ⇒ VersionDataCodec extra Term → Versions vNumber extra r → Peer (Handshake vNumber Term) AsClient StPropose m (Either (HandshakeClientProtocolError vNumber) r)

Handshake client which offers Versions vNumber vData to the remote peer.

TODO: GADT encoding of the client (Handshake.Client module).

Handshake server

handshakeServerPeer ∷ Ord vNumber ⇒ VersionDataCodec extra vParams → (∀ vData. extra vData → vData → vData → Accept) → Versions vNumber extra r → Peer (Handshake vNumber vParams) AsServer StPropose m (Either (RefuseReason vNumber) r)

Server following the handshake protocol; it accepts highest version offered by the peer that also belongs to the server versions.

TODO: GADT encoding of the server (Handshake.Server module).

Version data codec

data VersionDataCodec extra bites

Codec for version data (vData in code) exchanged by the handshake protocol.

Constructors

VersionDataCodec

Fields

encodeData ∷ ∀ vData. extra vData → vData → bites

encoder of vData which has access to 'extra vData' which can bring extra instances into the scope (by means of pattern matching on a GADT).

decodeData ∷ ∀ vData. extra vData → bites → Either Text vData

decoder of vData.

cborTermVersionDataCodec ∷ VersionDataCodec DictVersion Term

Tests

pureHandshake ∷ ∀ vNumber extra r. Ord vNumber ⇒ (∀ vData. extra vData → Dict Typeable vData) → (∀ vData. extra vData → vData → vData → Bool) → Versions vNumber extra r → Versions vNumber extra r → (Maybe r, Maybe r)

Pure computation which serves as a reference implementation of the Handshake protocol. Useful for testing handshakeClientPeer against handshakeServerPeer using connect