Ouroboros.Network.Protocol.Handshake.Version

Documentation

newtype Versions vNum extra r

The set of versions supported by the local agent are described by a map keyed on the version identifier.

If one needs to combine multiple versions the simplest way is to use the Semigroup instance.

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions ∷ Map vNum (Sigma (Version extra r))

newtype Application r vData

Takes a pair of version data: local then remote.

Constructors

Application

Fields

runApplication ∷ vData → vData → r

data Version extra r vData

Constructors

Version

Fields

versionApplication ∷ Application r vData

versionExtra ∷ extra vData

data Sigma f where

Constructors

Sigma ∷ !t → !(f t) → Sigma f

data Accept

A Maybe like type which better explains its purpose.

Constructors

Accept

Refuse !Text

class Acceptable v where

Methods

acceptableVersion ∷ v → v → Accept

data Dict constraint thing where

Constructors

Dict ∷ constraint thing ⇒ Dict constraint thing

data DictVersion vData where

Constructors

DictVersion ∷ (Typeable vData, Acceptable vData, Show vData) ⇒ CodecCBORTerm Text vData → DictVersion vData

data CodecCBORTerm fail a

Constructors

CodecCBORTerm

Fields

encodeTerm ∷ a → Term

decodeTerm ∷ Term → Either fail a

pickVersions ∷ Ord vNum ⇒ (∀ vData. extra vData → Dict Typeable vData) → Versions vNum extra r → Versions vNum extra r → Either (VersionMismatch vNum) (r, r)

Pick the version with the highest version number (by `Ord vNum`) common in both maps.

This is a useful guide for comparison with a version negotiation scheme for use in production between different processes. If the Versions maps used by each process are given to pickVersions, it should come up with the same result as the production version negotiation.

It is _assumed_ that if the maps agree on a key, then the existential types in the Sigma value at the key are also equal.

So, the issue here is that they may not have the same version data type. This becomes a non-issue on the network because the decoder/encoder basically fills the role of a safe dynamic type cast.

data VersionMismatch vNum where

Constructors

NoCommonVersion ∷ VersionMismatch vNum

InconsistentVersion ∷ vNum → VersionMismatch vNum

Simple or no versioning

simpleSingletonVersions ∷ vNum → vData → extra vData → r → Versions vNum extra r

Singleton smart constructor for Versions.

unversionedProtocol ∷ app → Versions UnversionedProtocol DictVersion app

Make a Versions for an unversioned protocol. Only use this for tests and demos where proper versioning is excessive.

foldMapVersions ∷ Ord vNum ⇒ (x → Versions vNum extra r) → NonEmpty x → Versions vNum extra r

Useful for folding a non-empty list of Versions.

A foldMap version for Semigroups which is restricted to Versions.

combineVersions ∷ Ord vNum ⇒ NonEmpty (Versions vNum extra r) → Versions vNum extra r

foldMapVersions'

Arguments

∷ Ord vNum

⇒ (x → Versions vNum extra r)

→ [x]
a non empty list of Versions

→ Versions vNum extra r

combineVersions'

Arguments

∷ Ord vNum

⇒ [Versions vNum extra r]
non empty list of Versions

→ Versions vNum extra r