Ouroboros.Network.Snocket

Snocket Interface

newtype Accept addr fd

Named pipes and Berkeley sockets have different API when accepting a connection. For named pipes the file descriptor created by createNamedPipe is supposed to be used for the first connected client. Named pipe accept loop looks this way:

acceptLoop k = do
  h <- createNamedPipe name
  connectNamedPipe h
  -- h is now in connected state
  forkIO (k h)
  acceptLoop k

For Berkeley sockets equivalent loop starts by creating a socket which accepts connections and accept returns a new socket in connected state

acceptLoop k = do
    s <- socket ...
    bind s address
    listen s
    loop s
  where
    loop s = do
      (s' , _addr') <- accept s
      -- s' is in connected state
      forkIO (k s')
      loop s

To make common API for both we use a recursive type Accept, see berkeleyAccept below. Creation of a socket / named pipe is part of Snocket, but this means we need to have different recursion step for named pipe & sockets. For sockets its recursion step will always return accept syscall; for named pipes the first callback will reuse the file descriptor created by open and only subsequent calls will create a new file descriptor by createNamedPipe, see namedPipeSnocket.

Constructors

Accept

Fields

runAccept ∷ IO (fd, addr, Accept addr fd)

fmapAccept ∷ (addr → addr') → Accept addr fd → Accept addr' fd

Arguments of Accept are in the wrong order.

TODO: this can be fixed later.

data AddressFamily addr where

We support either sockets or named pipes.

Constructors

SocketFamily ∷ !Family → AddressFamily SockAddr

LocalFamily ∷ AddressFamily LocalAddress

data Snocket m fd addr

Abstract communication interface that can be used by more than Socket. Snockets are polymorphic over monad which is used, this feature is useful for testing and/or simulations.

Constructors

Snocket

Fields

getLocalAddr ∷ fd → m addr

getRemoteAddr ∷ fd → m addr

addrFamily ∷ addr → AddressFamily addr

open ∷ AddressFamily addr → m fd

Open a file descriptor (socket / namedPipe). For named pipes this is using CreateNamedPipe syscall, for Berkeley sockets socket is used..

openToConnect ∷ addr → m fd

A way to create fd to pass to connect. For named pipes it will use CreateFile syscall. For Berkeley sockets this the same as open.

For named pipes we need full addr rather than just address family as it is for sockets.

connect ∷ fd → addr → m ()

connect is only needed for Berkeley sockets, for named pipes this is no-op.

bind ∷ fd → addr → m ()

listen ∷ fd → m ()

accept ∷ fd → Accept addr fd

close ∷ fd → m ()

toBearer ∷ Tracer m MuxTrace → fd → MuxBearer m

Socket based Snocktes

type SocketSnocket = Snocket IO Socket SockAddr

socketSnocket

Arguments

∷ IOManager
IOManager interface. We use it when we create a new socket and when we accept a connection.

Though it could be used in open, but that is going to be used in a bracket so it's better to keep it simple.

→ SocketSnocket

Create a Snocket for the given Family. In the bind method set ReuseAddr and ReusePort.

Local Snockets

type LocalSnocket = Snocket IO Socket LocalAddress

System dependent LocalSnocket type

localSnocket ∷ IOManager → FilePath → LocalSnocket

newtype LocalAddress

Local address, on Unix is associated with AF_UNIX family, on

Windows with `named-pipes`.

Constructors

LocalAddress

Fields

getFilePath ∷ FilePath

type LocalFD = Socket

localAddressFromPath ∷ FilePath → LocalAddress