Ouroboros.Network.Subscription.PeerState

Description

This module contains peer state management and error policies.

Documentation

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.

suspend ∷ Ord (Async m ()) ⇒ Maybe (PeerState m) → SuspendDecision Time → (Set (Async m ()), Maybe (PeerState m))

Action of SuspendDecision on Maybe PeerState. Action laws are only satisfied for the submonoid form by SuspendPeer and SuspendConsumer.

PeerStates and its operations

data PeerState m

Constructors

HotPeer !(Set (Async m ())) !(Set (Async m ()))
active peer with its producers and consumer threads

SuspendedConsumer !(Set (Async m ())) !Time
suspended consumer: with producer threads and time until the consumer is suspended

SuspendedPeer !Time !Time
suspended peer: producer & consumer suspend time

ColdPeer
peer with no opened connections in either direction

threadsToCancel ∷ Ord (Async m ()) ⇒ PeerState m → SuspendDecision diffTime → Set (Async m ())

Threads which needs to be cancelled when updating the PeerState with SuspendDecision.

data PeerStates m addr where

Map from addresses to PeerStates; it will be be shared in a StrictTVar.

Abstracting t is useful for tests, the IO version will use Time IO.

Constructors

PeerStates ∷ !(Map addr (PeerState m)) → PeerStates m addr
Map of peer states

ThrowException ∷ Exception e ⇒ e → PeerStates m addr
Or an exception to throw

newPeerStatesVar ∷ MonadSTM m ⇒ m (StrictTVar m (PeerStates m addr))

newPeerStatesVarSTM ∷ MonadSTM m ⇒ STM m (StrictTVar m (PeerStates m addr))

cleanPeerStates ∷ (MonadSTM m, MonadAsync m, MonadTime m, MonadTimer m) ⇒ DiffTime → StrictTVar m (PeerStates m addr) → m ()

Periodically clean PeerState. It will stop when PeerState becomes ThrowException.

runSuspendDecision ∷ ∀ m addr e. (Monad m, Ord addr, Ord (Async m ()), Exception e) ⇒ Time → addr → e → SuspendDecision DiffTime → PeerStates m addr → (PeerStates m addr, Set (Async m ()))

Update PeerStates for a given addr, using suspend, and return threads which must be cancelled.

This is more efficient that using the action of SuspendDecision on PeerStates, since it only uses a single dictionary lookup to update the state and return the set of threads to be cancelled.

registerConsumer ∷ ∀ m addr. (Ord addr, Ord (Async m ())) ⇒ addr → Async m () → PeerStates m addr → PeerStates m addr

Register consumer in PeerState. This is a partial function which assumes that the PeerState is HotPeer.

unregisterConsumer ∷ ∀ m addr. (Ord addr, Ord (Async m ())) ⇒ addr → Async m () → PeerStates m addr → PeerStates m addr

Unregister consumer from a PeerState.

registerProducer ∷ ∀ m addr. (Ord addr, Ord (Async m ())) ⇒ addr → Async m () → PeerStates m addr → PeerStates m addr

Register producer in PeerStates. This is a partial function which assumes that the PeerState is either HotPeer or SuspendedConsumer.

unregisterProducer ∷ ∀ m addr. (Ord addr, Ord (Async m ())) ⇒ addr → Async m () → PeerStates m addr → PeerStates m addr

type BeforeConnect m s addr = Time → addr → s → STM m (ConnectDecision s)

Check state before connecting to a remote peer. We will connect only if it retuns True.

data ConnectDecision s

Before connectin with a peer we make a decision to either connect to it or not.

Constructors

AllowConnection !s

DisallowConnection !s

runBeforeConnect ∷ (MonadSTM m, MonadTime m) ⇒ StrictTVar m s → BeforeConnect m s addr → addr → m Bool

Run BeforeConnect callback in a MonadTime monad.

beforeConnectTx ∷ ∀ m addr. (MonadSTM m, Ord addr) ⇒ BeforeConnect m (PeerStates m addr) addr

BeforeConnect callback: it updates peer state and return boolean value wheather to connect to it or not. If a peer hasn't been recorded in PeerStates, we add it and try to connect to it.

Re-exports

data DiffTime

This is a length of time, as measured by a clock. Conversion functions will treat it as seconds. It has a precision of 10^-12 s.

Auxiliary functions

alterAndLookup ∷ ∀ k s a. Ord k ⇒ (Maybe a → (s, Maybe a)) → k → Map k a → (Map k a, Maybe s)